refactor(agent): split data tools into tools/data with shims
Move DuckDB/query/chart/dashboard helpers under helpudoc_agent.tools.data,
keep data_agent_tools and data_report_renderers as compatibility entrypoints,
and resolve monkeypatch targets across agent/helpudoc_agent import aliases.
