+++ agent/helpudoc_agent/data_agent_tools.py
@@ -1,2852 +1,35 @@
-import ast
-import base64
-import hashlib
-import html
-import json
-import logging
-import multiprocessing as mp
-import os
-import re
-import shutil
-import threading
-from dataclasses import dataclass
-from datetime import date, datetime, timedelta, timezone
-from pathlib import Path
-from queue import Empty
-from typing import Annotated, Any, Dict, List, Optional, Set, Tuple
-from uuid import uuid4
+"""DuckDB-backed data tools — compatibility shim for ``helpudoc_agent.data_agent_tools``.
 
-import duckdb
-import numpy as np
-import pandas as pd
-from langchain_core.callbacks import CallbackManagerForToolRun
-from langchain_core.tools import Tool, tool
-from pydantic import Field
+Implementations live under ``helpudoc_agent.tools.data``. This module re-exports
+symbols so existing imports, skill entrypoints, and test monkeypatch paths keep
+working (for example ``run_bigquery_query`` and ``CHART_EXECUTION_TIMEOUT_SECONDS``).
+"""
+from __future__ import annotations
 
-from .bigquery_export_tools import (
-    extract_bearer_header,
-    load_bigquery_toolbox_config,
-    resolve_output_path,
-    run_bigquery_query,
-    validate_read_only_sql,
-    write_export_dataframe,
+from helpudoc_agent.bigquery_export_tools import run_bigquery_query
+from helpudoc_agent.tools.data.constants import (
+    CHART_EXECUTION_TIMEOUT_SECONDS,
+    MAX_CHART_COUNT,
+    MAX_QUERY_COUNT,
 )
-from .data_report_renderers import render_dashboard_html, render_summary_html
-from .state import WorkspaceState
-from .tagged_file_policy import tagged_files_mode_guard
-
-logger = logging.getLogger(__name__)
-
-# Optional plotting libraries (used inside the chart tool)
-try:  # pragma: no cover - environment dependent
-    import matplotlib
-
-    matplotlib.use("Agg")
-    import matplotlib.pyplot as plt
-except Exception:  # pragma: no cover - optional dependency
-    plt = None
-    logger.warning("matplotlib not available; chart images may fail")
-
-try:  # pragma: no cover - environment dependent
-    import seaborn as sns
-except Exception:  # pragma: no cover - optional dependency
-    sns = None
-    logger.warning("seaborn not available; chart images may fail")
-
-ALLOWED_ARTIFACT_EXTENSIONS = {
-    ".json": "application/json",
-    ".html": "text/html",
-    ".htm": "text/html",
-    ".png": "image/png",
-    ".jpg": "image/jpeg",
-    ".jpeg": "image/jpeg",
-    ".csv": "text/csv",
-    ".txt": "text/plain",
-    ".md": "text/markdown",
-    ".parquet": "application/octet-stream",
-}
-MAX_RESULT_ROWS = 20
-MAX_SESSION_ROWS = 1000
-MAX_QUERY_COUNT = 10
-MAX_CHART_COUNT = 5
-STRICT_DASHBOARD_QUERY_COUNT = 5
-STRICT_DASHBOARD_PREVIEW_QUERY_COUNT = 1
-STRICT_DASHBOARD_SCHEMA_COUNT = 1
-STRICT_DASHBOARD_MIN_CHART_COUNT = 3
-DEFAULT_CACHE_TTL_HOURS = 24
-MAX_MATERIALIZED_ROWS = 100000
-MAX_QUERY_RESULT_ROWS = MAX_SESSION_ROWS + 1
-CHART_EXECUTION_TIMEOUT_SECONDS = 5.0
-NATIVE_DASHBOARD_CHART_TYPES = {"bar", "line", "scatter", "pie", "area"}
-NATIVE_DASHBOARD_AGGREGATIONS = {"count", "sum", "avg", "mean", "min", "max", "nunique", "count_distinct"}
-NATIVE_DASHBOARD_SORT_FIELDS = {"x", "y"}
-NATIVE_DASHBOARD_SORT_DIRECTIONS = {"asc", "desc"}
-NATIVE_DASHBOARD_ORIENTATIONS = {"", "h", "v"}
-WORKSPACE_SCAN_EXCLUDED_DIRS = {
-    ".git",
-    ".hg",
-    ".svn",
-    ".venv",
-    "__pycache__",
-    "node_modules",
-    "dist",
-    "build",
-    ".mypy_cache",
-    ".pytest_cache",
-    ".ruff_cache",
-    ".tox",
-    ".idea",
-    ".vscode",
-}
-DATA_FILE_EXTENSIONS = {".csv", ".parquet"}
-STRICT_DASHBOARD_DIMENSION_FIELDS = (
-    "country",
-    "device_type",
-    "browser",
-    "traffic_source",
-    "category",
-    "product_category",
-    "age_group",
-)
-DATA_DISCOVERY_DIR_CANDIDATES = (
-    "data",
-    "datasets",
-    "exports",
-    "data_exports",
-)
-
-
-@dataclass
-class _QueryRecord:
-    sql: str
-    row_count: int
-    preview: "pd.DataFrame"
-    truncated: bool = False
-
-
-@dataclass
-class _ChartRecord:
-    title: str
-    artifact_paths: List[str]
-
-
-@dataclass
-class _MaterializationRecord:
-    cache_key: str
-    sql: str
-    parquet_path: str
-    metadata_path: str
-    row_count: int
-    connector: str
-    cached: bool
-    expires_at: str
-
-
-class DataAgentSessionState:
-    """Holds per-run guardrails and history for the data agent tools."""
-
-    def __init__(self) -> None:
-        self.reset()
-
-    def reset(self) -> None:
-        self.schema_inspected = False
-        self.schema_read_count = 0
-        self.query_count = 0
-        self.chart_count = 0
-        self.summary_generated = False
-        self.dashboard_generated = False
-        self.last_query_result: Optional[pd.DataFrame] = None
-        self.last_query_sql: Optional[str] = None
-        self.last_schema_result: Optional[str] = None
-        self.query_history: List[_QueryRecord] = []
-        self.chart_history: List[_ChartRecord] = []
-        self.materialization_history: List[_MaterializationRecord] = []
-        self.run_artifacts: List[Dict[str, Any]] = []
-        self.dashboard_dimension_signatures: Set[str] = set()
-
-
-class SafePandasProxy:
-    """Restricts pandas entry points that could be used to read raw files."""
-
-    _blocked = {"read_csv", "read_parquet", "read_json", "read_excel"}
-
-    def __init__(self, module: Any):
-        self._module = module
-
-    def __getattr__(self, name: str) -> Any:
-        if name in self._blocked:
-            raise ValueError(
-                "Direct file reads are disabled in this environment. "
-                "Use DuckDB queries to load data."
-            )
-        return getattr(self._module, name)
-
-
-def _build_safe_builtins() -> Dict[str, Any]:
-    return {
-        "abs": abs,
-        "all": all,
-        "any": any,
-        "min": min,
-        "max": max,
-        "sum": sum,
-        "len": len,
-        "round": round,
-        "range": range,
-        "enumerate": enumerate,
-        "zip": zip,
-        "sorted": sorted,
-        "bool": bool,
-        "int": int,
-        "float": float,
-        "str": str,
-        "list": list,
-        "dict": dict,
-        "tuple": tuple,
-        "set": set,
-        "isinstance": isinstance,
-        "type": type,
-        "__import__": __import__,
-    }
-
-
-def _utc_now() -> datetime:
-    return datetime.now(timezone.utc)
-
-
-def _workspace_rel(path: Path, root: Path) -> str:
-    return path.resolve().relative_to(root.resolve()).as_posix()
-
-
-def _safe_slug(value: str, default: str) -> str:
-    slug = re.sub(r"[^a-zA-Z0-9_-]+", "_", (value or "").strip()).strip("_")
-    return slug or default
-
-
-def _should_skip_directory(name: str) -> bool:
-    if name in WORKSPACE_SCAN_EXCLUDED_DIRS:
-        return True
-    return name.startswith(".")
-
-
-def _iter_workspace_files(
-    root: Path,
-    *,
-    allowed_extensions: Optional[Set[str]] = None,
-    preferred_dirs: Tuple[str, ...] = (),
-    include_root_recursive: bool = False,
-) -> List[Path]:
-    seen: Set[Path] = set()
-    files: List[Path] = []
-
-    def _append_file(path: Path) -> None:
-        if allowed_extensions and path.suffix.lower() not in allowed_extensions:
-            return
-        resolved = path.resolve()
-        if resolved in seen:
-            return
-        seen.add(resolved)
-        files.append(path)
-
-    for child in root.iterdir():
-        if child.is_file():
-            _append_file(child)
-
-    preferred_roots: List[Path] = []
-    for dirname in preferred_dirs:
-        candidate = root / dirname
-        if candidate.exists() and candidate.is_dir():
-            preferred_roots.append(candidate)
-
-    def _scan_recursive(base: Path) -> None:
-        for current_root, dirnames, filenames in os.walk(base):
-            current_path = Path(current_root)
-            if current_path != base and _should_skip_directory(current_path.name):
-                dirnames[:] = []
-                continue
-            dirnames[:] = [
-                dirname for dirname in dirnames if not _should_skip_directory(dirname)
-            ]
-            for filename in filenames:
-                _append_file(current_path / filename)
-
-    for base in preferred_roots:
-        _scan_recursive(base)
-    if include_root_recursive or not preferred_roots:
-        _scan_recursive(root)
-    return files
-
-
-def _snapshot_workspace(root: Path) -> Dict[str, Tuple[int, int]]:
-    snapshot: Dict[str, Tuple[int, int]] = {}
-    for path in _iter_workspace_files(
-        root,
-        allowed_extensions=set(ALLOWED_ARTIFACT_EXTENSIONS.keys()),
-        preferred_dirs=("charts", "reports", "dashboards", "exports", "data_exports"),
-        include_root_recursive=True,
-    ):
-        rel = path.relative_to(root).as_posix()
-        try:
-            stat = path.stat()
-        except OSError:
-            continue
-        snapshot[rel] = (int(stat.st_mtime * 1e9), stat.st_size)
-    return snapshot
-
-
-def _detect_new_files(
-    root: Path, before: Dict[str, Tuple[int, int]], after: Dict[str, Tuple[int, int]]
-) -> List[Dict[str, Any]]:
-    artifacts: List[Dict[str, Any]] = []
-    for rel, meta in after.items():
-        if rel in before:
-            continue
-        path = root / rel
-        ext = path.suffix.lower()
-        if ext not in ALLOWED_ARTIFACT_EXTENSIONS:
-            continue
-        mime = ALLOWED_ARTIFACT_EXTENSIONS[ext]
-        artifacts.append(
-            {
-                "path": rel,
-                "mimeType": mime,
-                "size": meta[1],
-            }
-        )
-    return artifacts
-
-
-def _cleanup_new_files(
-    root: Path,
-    before: Dict[str, Tuple[int, int]],
-    after: Dict[str, Tuple[int, int]],
-) -> None:
-    for rel in after:
-        if rel in before:
-            continue
-        path = root / rel
-        try:
-            if path.exists():
-                path.unlink()
-        except OSError:
-            logger.warning("Failed to clean up artifact %s", path, exc_info=True)
-
-
-def _format_dataframe_markdown(df: pd.DataFrame, *, truncated: bool = False) -> str:
-    if df.empty:
-        return "Query executed successfully but returned no data."
-
-    display_df = df.head(MAX_RESULT_ROWS)
-    message_lines = [f"Result shape: {len(df)} rows x {len(df.columns)} columns."]
-    if truncated:
-        message_lines.append(
-            f"Execution was safety-capped at {MAX_SESSION_ROWS} rows. Refine the query for the full result."
-        )
-    if len(df.columns):
-        columns = ", ".join(f"`{column}`" for column in df.columns[:10])
-        if len(df.columns) > 10:
-            columns += ", ..."
-        message_lines.append(f"Columns: {columns}")
-    numeric_summary = _format_numeric_summary(df)
-    if numeric_summary:
-        message_lines.append(f"Numeric summary: {numeric_summary}")
-    if len(df) > MAX_RESULT_ROWS:
-        message_lines.append(f"Showing the first {MAX_RESULT_ROWS} rows below.")
-    message_lines.append(display_df.to_markdown())
-    rendered = "\n".join(message_lines)
-    if len(rendered) > 4000:
-        return rendered[:4000] + "\n... (Output truncated due to length)"
-    return rendered
-
-
-def _format_sample_value(value: Any) -> str:
-    if isinstance(value, np.ndarray):
-        value = value.tolist()
-    if isinstance(value, pd.Series):
-        value = value.tolist()
-    if isinstance(value, (list, tuple, set, dict)):
-        structured = json.dumps(value, default=str)
-        return structured if len(structured) <= 32 else structured[:29] + "..."
-    try:
-        is_null = pd.isna(value)
-    except (TypeError, ValueError):
-        is_null = False
-    if is_null:
-        return "null"
-    if isinstance(value, str):
-        compact = value if len(value) <= 32 else value[:29] + "..."
-        return repr(compact)
-    if isinstance(value, (datetime, pd.Timestamp)):
-        return value.isoformat()
-    return str(value)
-
-
-def _format_numeric_summary(df: pd.DataFrame) -> str:
-    numeric_df = df.select_dtypes(include=["number"])
-    if numeric_df.empty:
-        return ""
-
-    summaries: List[str] = []
-    for column in numeric_df.columns[:3]:
-        series = numeric_df[column].dropna()
-        if series.empty:
-            continue
-        summaries.append(
-            f"`{column}` min={series.min():.3g}, median={series.median():.3g}, max={series.max():.3g}"
-        )
-    return "; ".join(summaries)
-
-
-def _query_looks_aggregated(query: str) -> bool:
-    if re.search(r"\bover\s*\(", query, re.IGNORECASE):
-        return False
-    return bool(
-        re.search(r"\bgroup\s+by\b", query, re.IGNORECASE)
-        or re.search(r"\bhaving\b", query, re.IGNORECASE)
-    )
-
-
-def _get_dashboard_mode(workspace_state: WorkspaceState) -> Dict[str, Any]:
-    context = getattr(workspace_state, "context", {}) or {}
-    raw = context.get("dashboard_mode")
-    if isinstance(raw, dict):
-        return raw
-    active_skill = str(context.get("active_skill") or "").strip()
-    tagged_files = context.get("tagged_files") or []
-    tagged_dataset_paths = [
-        str(path).strip()
-        for path in tagged_files
-        if str(path).strip() and Path(str(path).strip()).suffix.lower() in DATA_FILE_EXTENSIONS
-    ]
-    return {
-        "strictLocalDatasets": active_skill == "data/dashboard" and bool(tagged_dataset_paths),
-        "taggedDatasetPaths": tagged_dataset_paths,
-    }
-
-
-def _is_strict_dashboard_mode(workspace_state: WorkspaceState) -> bool:
-    mode = _get_dashboard_mode(workspace_state)
-    return bool(mode.get("strictLocalDatasets"))
-
-
-def _is_plan_approved(workspace_state: WorkspaceState) -> bool:
-    context = getattr(workspace_state, "context", {}) or {}
-    if context.get("skip_plan_approvals"):
-        return True
-    return bool(context.get("plan_approved"))
-
-
-def _dashboard_plan_gate_message() -> str:
-    return (
-        "Dashboard planning mode is active. Draft the dashboard plan, call request_plan_approval, "
-        "and wait for approval before running aggregate analysis or generating charts."
-    )
-
-
-def _looks_like_preview_query(query: str) -> bool:
-    if _query_looks_aggregated(query):
-        return False
-    limit_match = re.search(r"\blimit\s+(\d+)\b", query, re.IGNORECASE)
-    if limit_match:
-        try:
-            return int(limit_match.group(1)) <= 50
-        except ValueError:
-            return False
-    return bool(re.search(r"\bselect\s+\*\s+from\b", query, re.IGNORECASE))
-
-
-def _extract_dashboard_dimension_signature(query: str) -> Optional[str]:
-    if not _query_looks_aggregated(query):
-        return None
-    lower = query.lower()
-    matched = [
-        field
-        for field in STRICT_DASHBOARD_DIMENSION_FIELDS
-        if re.search(rf"\b{re.escape(field)}\b", lower)
-    ]
-    if not matched:
-        return None
-    return "|".join(sorted(set(matched)))
-
-
-def _json_safe_chart_value(value: Any) -> Any:
-    if value is None:
-        return None
-    try:
-        json.dumps(value)
-        return value
-    except TypeError:
-        return str(value)
-
-
-def _chart_payload(
-    chart_title: str,
-    *,
-    status: str,
-    message: str,
-    row_count: int,
-    chart_config: Any = None,
-    plotly_config_path: Optional[str] = None,
-    output_files: Optional[List[Dict[str, Any]]] = None,
-    error_type: Optional[str] = None,
-    available_columns: Optional[List[str]] = None,
-) -> str:
-    payload: Dict[str, Any] = {
-        "status": status,
-        "message": message,
-        "chart_title": chart_title,
-        "chart_config": _json_safe_chart_value(chart_config),
-        "plotly_config_path": plotly_config_path,
-        "plotly_html_path": None,
-        "output_files": output_files or [],
-        "row_count": row_count,
-    }
-    if error_type:
-        payload["error_type"] = error_type
-    if available_columns is not None:
-        payload["available_columns"] = available_columns
-    return _json_dump(payload)
-
-
-def _execute_chart_code_worker(
-    result_queue: "mp.Queue[Dict[str, Any]]",
-    sanitized_code: str,
-    chart_title: str,
-    df_context: pd.DataFrame,
-) -> None:
-    safe_globals: Dict[str, Any] = {
-        "__builtins__": _build_safe_builtins(),
-        "pd": SafePandasProxy(pd),
-        "np": np,
-        "json": json,
-        "isinstance": isinstance,
-        "len": len,
-        "range": range,
-        "enumerate": enumerate,
-        "chart_config": None,
-        "chart_title": chart_title,
-        "df": df_context,
-    }
-    if plt is not None:
-        safe_globals["plt"] = plt
-    if sns is not None:
-        safe_globals["sns"] = sns
-
-    exec_namespace: Dict[str, Any] = {}
-    try:
-        exec(sanitized_code, safe_globals, exec_namespace)
-    except KeyError as exc:
-        result_queue.put(
-            {
-                "status": "error",
-                "error_type": "missing_column",
-                "message": f"Column {str(exc)} not found in query result.",
-                "available_columns": list(df_context.columns),
-            }
-        )
-        return
-    except Exception as exc:
-        result_queue.put(
-            {
-                "status": "error",
-                "error_type": "execution_error",
-                "message": f"Failed to execute chart code: {exc}",
-            }
-        )
-        return
-
-    chart_config = exec_namespace.get("chart_config") or safe_globals.get("chart_config")
-    plotly_payload = _coerce_plotly_spec(chart_config)
-    if plotly_payload is None:
-        plotly_payload = _coerce_plotly_spec(
-            exec_namespace.get("plotly_fig") or exec_namespace.get("fig")
-        )
-
-    result_queue.put(
-        {
-            "status": "success",
-            "chart_config": _json_safe_chart_value(chart_config),
-            "plotly_payload": plotly_payload,
-        }
-    )
-
-
-def _run_chart_code_in_subprocess(
-    sanitized_code: str,
-    chart_title: str,
-    df_context: pd.DataFrame,
-    *,
-    timeout_seconds: Optional[float] = None,
-) -> Dict[str, Any]:
-    effective_timeout = timeout_seconds or CHART_EXECUTION_TIMEOUT_SECONDS
-    ctx = mp.get_context("spawn")
-    result_queue: "mp.Queue[Dict[str, Any]]" = ctx.Queue()
-    try:
-        process = ctx.Process(
-            target=_execute_chart_code_worker,
-            args=(result_queue, sanitized_code, chart_title, df_context),
-        )
-        process.start()
-        process.join(effective_timeout)
-
-        if process.is_alive():
-            process.terminate()
-            process.join(1)
-            if process.is_alive():
-                process.kill()
-                process.join(1)
-            return {
-                "status": "error",
-                "error_type": "timeout",
-                "message": f"Chart execution timed out after {effective_timeout:g} seconds.",
-            }
-
-        try:
-            result = result_queue.get_nowait()
-        except Empty:
-            result = None
-        if result is None:
-            return {
-                "status": "error",
-                "error_type": "execution_error",
-                "message": "Chart subprocess exited without returning a result.",
-            }
-        return result
-    finally:
-        result_queue.close()
-        result_queue.join_thread()
-
-
-def _markdown_to_html(markdown_text: str) -> str:
-    if not markdown_text:
-        return ""
-
-    text = markdown_text.replace("\r\n", "\n").replace("\r", "\n")
-    text = html.escape(text)
-
-    code_blocks: List[Tuple[str, str]] = []
-
-    def _capture_code_block(match: re.Match[str]) -> str:
-        lang = match.group(1) or ""
-        code = match.group(2) or ""
-        idx = len(code_blocks)
-        code_blocks.append((lang, code))
-        return f"@@CODEBLOCK{idx}@@"
-
-    text = re.sub(r"```([\w+-]*)\n([\s\S]*?)\n```", _capture_code_block, text)
-
-    text = re.sub(r"`([^`]+)`", r"<code>\1</code>", text)
-    text = re.sub(r"\*\*(.+?)\*\*", r"<strong>\1</strong>", text)
-
-    for level in range(6, 0, -1):
-        pattern = rf"^{'#' * level}\s+(.*)$"
-        text = re.sub(pattern, rf"<h{level}>\1</h{level}>", text, flags=re.M)
-
-    text = re.sub(r"^---+$", "<hr />", text, flags=re.M)
-
-    lines = text.split("\n")
-    output: List[str] = []
-    in_list = False
-    for line in lines:
-        match = re.match(r"^\s*[-*]\s+(.*)$", line)
-        if match:
-            if not in_list:
-                output.append("<ul>")
-                in_list = True
-            output.append(f"<li>{match.group(1)}</li>")
-        else:
-            if in_list:
-                output.append("</ul>")
-                in_list = False
-            output.append(line)
-    if in_list:
-        output.append("</ul>")
-
-    text = "\n".join(output)
-
-    for idx, (lang, code) in enumerate(code_blocks):
-        class_attr = f" class=\"language-{lang}\"" if lang else ""
-        code_html = f"<pre><code{class_attr}>{code}</code></pre>"
-        text = text.replace(f"@@CODEBLOCK{idx}@@", code_html)
-
-    blocks = re.split(r"\n{2,}", text.strip())
-    wrapped: List[str] = []
-    block_start = re.compile(r"^(<h\d|<ul>|<ol>|<pre>|<hr\s*/?>)")
-    for block in blocks:
-        stripped = block.strip()
-        if not stripped:
-            continue
-        if block_start.match(stripped):
-            wrapped.append(stripped)
-        else:
-            wrapped.append(f"<p>{stripped}</p>")
-    return "\n".join(wrapped)
-
-
-def _chart_title_from_path(path: Path) -> str:
-    title = path.stem
-    if title.endswith(".plotly"):
-        title = title[: -len(".plotly")]
-    return title.replace("_", " ")
-
-
-def _json_default(value: Any) -> Any:
-    if isinstance(value, (np.integer,)):
-        return int(value)
-    if isinstance(value, (np.floating,)):
-        return float(value)
-    if isinstance(value, (np.ndarray,)):
-        return value.tolist()
-    if isinstance(value, (pd.Series, pd.Index)):
-        return value.tolist()
-    if isinstance(value, pd.Timestamp):
-        return value.isoformat()
-    if isinstance(value, Path):
-        return value.as_posix()
-    if isinstance(value, datetime):
-        return value.isoformat()
-    raise TypeError(f"Object of type {type(value)} is not JSON serializable")
-
-
-def _json_dump(payload: Any) -> str:
-    return json.dumps(payload, default=_json_default, ensure_ascii=False)
-
-
-def _coerce_plotly_spec(payload: Any) -> Optional[Dict[str, Any]]:
-    """Attempt to coerce a payload into a Plotly JSON-serializable dict."""
-    if payload is None:
-        return None
-
-    try:  # pragma: no cover - optional dependency
-        import plotly.io as pio  # type: ignore
-    except Exception:  # pragma: no cover - optional dependency
-        pio = None
-
-    try:
-        if hasattr(payload, "to_plotly_json"):
-            if pio is not None:
-                return json.loads(pio.to_json(payload, validate=False))
-            return payload.to_plotly_json()
-        if isinstance(payload, dict):
-            data = payload.get("data")
-            if isinstance(data, list):
-                return payload
-    except Exception:  # pragma: no cover - defensive
-        logger.warning("Failed to coerce Plotly spec", exc_info=True)
-    return None
-
-
-def _coerce_text_arg(value: Any, default: str = "") -> str:
-    if value is None:
-        return default
-    if isinstance(value, str):
-        return value
-    return str(value)
-
-
-def _coerce_int_arg(
-    value: Any,
-    default: int,
-    *,
-    minimum: Optional[int] = None,
-    maximum: Optional[int] = None,
-) -> int:
-    try:
-        coerced = int(value if value is not None else default)
-    except (TypeError, ValueError):
-        coerced = int(default)
-    if minimum is not None:
-        coerced = max(minimum, coerced)
-    if maximum is not None:
-        coerced = min(maximum, coerced)
-    return coerced
-
-
-def _coerce_bool_arg(value: Any, default: bool = False) -> bool:
-    if isinstance(value, bool):
-        return value
-    if value is None:
-        return default
-    if isinstance(value, str):
-        normalized = value.strip().lower()
-        if normalized in {"1", "true", "yes", "y", "on"}:
-            return True
-        if normalized in {"0", "false", "no", "n", "off"}:
-            return False
-    return bool(default if value is None else value)
-
-
-def _resolve_workspace_html_path(
-    workspace_root: Path,
-    output_path: str,
-    *,
-    default_dir: str,
-    default_stem: str,
-) -> Path:
-    raw = (output_path or "").strip()
-    if not raw:
-        raw = f"{default_dir}/{default_stem}.html"
-    elif raw.endswith("/"):
-        raw = f"{raw}{default_stem}.html"
-    elif Path(raw).suffix.lower() not in {".html", ".htm"}:
-        raw = f"{raw}.html"
-
-    candidate = Path(raw)
-    if candidate.is_absolute():
-        candidate = Path(str(candidate).lstrip("/"))
-    resolved = (workspace_root / candidate).resolve()
-    root_resolved = workspace_root.resolve()
-    if root_resolved not in resolved.parents and resolved != root_resolved:
-        raise ValueError("output_path must remain inside the workspace.")
-    return resolved
-
-
-def _resolve_workspace_dashboard_dir(
-    workspace_root: Path,
-    output_path: str,
-    *,
-    default_dir: str,
-    default_stem: str,
-) -> Path:
-    raw = (output_path or "").strip()
-    if not raw:
-        raw = f"{default_dir}/{default_stem}"
-    elif raw.endswith("/"):
-        raw = raw.rstrip("/")
-    else:
-        candidate = Path(raw)
-        suffix = candidate.suffix.lower()
-        if suffix in {".html", ".htm"}:
-            raw = str(candidate.with_suffix(""))
-
-    candidate = Path(raw)
-    if candidate.is_absolute():
-        candidate = Path(str(candidate).lstrip("/"))
-    resolved = (workspace_root / candidate).resolve()
-    root_resolved = workspace_root.resolve()
-    if root_resolved not in resolved.parents and resolved != root_resolved:
-        raise ValueError("output_path must remain inside the workspace.")
-    return resolved
-
-
-def _resolve_workspace_data_path(workspace_root: Path, raw_path: str) -> Path:
-    candidate = Path((raw_path or "").strip())
-    if not str(candidate):
-        raise ValueError("dashboard_dataset_path is required for data-backed filtering.")
-    if candidate.is_absolute():
-        candidate = Path(str(candidate).lstrip("/"))
-    resolved = (workspace_root / candidate).resolve()
-    root_resolved = workspace_root.resolve()
-    if root_resolved not in resolved.parents and resolved != root_resolved:
-        raise ValueError("dashboard_dataset_path must remain inside the workspace.")
-    return resolved
-
-
-def _normalize_record_value(value: Any) -> Any:
-    if value is None or pd.isna(value):
-        return None
-    if isinstance(value, (pd.Timestamp, datetime, date)):
-        return value.isoformat()
-    if isinstance(value, np.datetime64):
-        return pd.Timestamp(value).isoformat()
-    if isinstance(value, np.generic):
-        return value.item()
-    return value
-
-
-def _load_dashboard_dataset(workspace_root: Path, dataset_path: str) -> Tuple[List[Dict[str, Any]], List[Dict[str, Any]], str]:
-    resolved = _resolve_workspace_data_path(workspace_root, dataset_path)
-    if not resolved.exists():
-        raise ValueError(f"Dashboard dataset not found: {dataset_path}")
-
-    suffix = resolved.suffix.lower()
-    if suffix == ".parquet":
-        df = pd.read_parquet(resolved)
-        format_name = "parquet"
-    elif suffix == ".csv":
-        df = pd.read_csv(resolved)
-        format_name = "csv"
-    elif suffix in {".json", ".jsonl", ".ndjson"}:
-        format_name = "json"
-        if suffix == ".json":
-            payload = json.loads(resolved.read_text(encoding="utf-8"))
-            if isinstance(payload, dict):
-                if isinstance(payload.get("rows"), list):
-                    rows = payload.get("rows") or []
-                elif isinstance(payload.get("data"), list):
-                    rows = payload.get("data") or []
-                else:
-                    rows = [payload]
-            elif isinstance(payload, list):
-                rows = payload
-            else:
-                raise ValueError("JSON dashboard datasets must contain an object or list.")
-        else:
-            rows = [
-                json.loads(line)
-                for line in resolved.read_text(encoding="utf-8").splitlines()
-                if line.strip()
-            ]
-        df = pd.DataFrame(rows)
-    else:
-        raise ValueError(
-            "dashboard_dataset_path must point to a .parquet, .csv, .json, .jsonl, or .ndjson file."
-        )
-
-    records = [
-        {str(key): _normalize_record_value(value) for key, value in row.items()}
-        for row in df.replace({np.nan: None}).to_dict(orient="records")
-    ]
-    schema = [
-        {
-            "name": str(column),
-            "type": str(dtype),
-        }
-        for column, dtype in df.dtypes.items()
-    ]
-    return records, schema, format_name
-
-
-def _build_dataset_reference(workspace_root: Path, dataset_path: str) -> Dict[str, Any]:
-    cleaned = _coerce_text_arg(dataset_path).strip()
-    if not cleaned:
-        return {}
-    resolved = _resolve_workspace_data_path(workspace_root, cleaned)
-    if not resolved.exists():
-        return {"path": cleaned}
-    stat = resolved.stat()
-    return {
-        "path": _workspace_rel(resolved, workspace_root),
-        "format": resolved.suffix.lower().lstrip("."),
-        "size": stat.st_size,
-        "updatedAt": datetime.fromtimestamp(stat.st_mtime, tz=timezone.utc).isoformat(),
-        "version": f"{int(stat.st_mtime)}-{stat.st_size}",
-    }
-
-
-def _build_dashboard_dataset_block(
-    workspace_root: Path,
-    dashboard_dataset_path: str,
-    dataset_records: List[Dict[str, Any]],
-    preview_rel_path: str,
-) -> Dict[str, Any]:
-    """Canonical dataset path plus browser-friendly preview rows (v1 JSON)."""
-    base = _build_dataset_reference(workspace_root, dashboard_dataset_path)
-    if not dataset_records:
-        return base
-    block = dict(base)
-    block["previewPath"] = preview_rel_path
-    block["rowCount"] = len(dataset_records)
-    return block
-
-
-def _normalize_filter_schema(raw_filters: Optional[List[Dict[str, Any]]]) -> List[Dict[str, Any]]:
-    normalized: List[Dict[str, Any]] = []
-    for idx, raw in enumerate(raw_filters or [], start=1):
-        if not isinstance(raw, dict):
-            continue
-        field_name = _coerce_text_arg(raw.get("field")).strip()
-        filter_type = _coerce_text_arg(raw.get("type"), "categorical").strip().lower()
-        if not field_name:
-            continue
-        normalized.append(
-            {
-                "id": _coerce_text_arg(raw.get("id"), f"filter_{idx}").strip() or f"filter_{idx}",
-                "field": field_name,
-                "label": _coerce_text_arg(raw.get("label"), field_name.replace("_", " ").title()).strip()
-                or field_name.replace("_", " ").title(),
-                "type": filter_type if filter_type in {"categorical", "date", "datetime", "numeric"} else "categorical",
-                "operators": list(raw.get("operators") or []),
-                "multi": _coerce_bool_arg(raw.get("multi"), True),
-                "options": list(raw.get("options") or []),
-                "applies_to": list(raw.get("applies_to") or []),
-                "default": raw.get("default"),
-                "presets": list(raw.get("presets") or []),
-                "step": raw.get("step"),
-            }
-        )
-    return normalized
-
-
-def _normalize_chart_bindings(raw_bindings: Optional[List[Dict[str, Any]]]) -> Dict[int, Dict[str, Any]]:
-    normalized: Dict[int, Dict[str, Any]] = {}
-    for idx, raw in enumerate(raw_bindings or [], start=1):
-        if not isinstance(raw, dict):
-            continue
-        chart_index = _coerce_int_arg(raw.get("chart_index"), idx, minimum=1)
-        normalized[chart_index] = {
-            "chart_index": chart_index,
-            "chart_type": _coerce_text_arg(raw.get("chart_type"), "bar").strip().lower() or "bar",
-            "x_field": _coerce_text_arg(raw.get("x_field")).strip(),
-            "y_field": _coerce_text_arg(raw.get("y_field")).strip(),
-            "dimension_field": _coerce_text_arg(raw.get("dimension_field")).strip(),
-            "metric_field": _coerce_text_arg(raw.get("metric_field")).strip(),
-            "aggregation": _coerce_text_arg(raw.get("aggregation"), "count").strip().lower() or "count",
-            "numerator_field": _coerce_text_arg(raw.get("numerator_field")).strip(),
-            "denominator_field": _coerce_text_arg(raw.get("denominator_field")).strip(),
-            "series_field": _coerce_text_arg(raw.get("series_field")).strip(),
-            "orientation": _coerce_text_arg(raw.get("orientation")).strip().lower(),
-            "sort_by": _coerce_text_arg(raw.get("sort_by"), "y").strip().lower() or "y",
-            "sort_direction": _coerce_text_arg(raw.get("sort_direction"), "desc").strip().lower() or "desc",
-            "limit": _coerce_int_arg(raw.get("limit"), 0, minimum=0),
-            "mode": _coerce_text_arg(raw.get("mode"), "lines+markers").strip(),
-            "title": _coerce_text_arg(raw.get("title")).strip(),
-            "x_title": _coerce_text_arg(raw.get("x_title")).strip(),
-            "y_title": _coerce_text_arg(raw.get("y_title")).strip(),
-            "question_answered": _coerce_text_arg(raw.get("question_answered")).strip(),
-            "why_it_matters": _coerce_text_arg(raw.get("why_it_matters")).strip(),
-            "highlight_rule": _coerce_text_arg(raw.get("highlight_rule")).strip(),
-            "format": _coerce_text_arg(raw.get("format")).strip(),
-            "interactive": _coerce_bool_arg(raw.get("interactive"), True),
-            "layout_span": _coerce_text_arg(raw.get("layout_span"), "half").strip().lower() or "half",
-            "layout_section": _coerce_text_arg(raw.get("layout_section"), "drivers").strip().lower() or "drivers",
-            "static_reason": _coerce_text_arg(raw.get("static_reason")).strip(),
-        }
-    return normalized
-
-
-def _schema_type_map(dataset_schema: List[Dict[str, Any]]) -> Dict[str, str]:
-    return {
-        _coerce_text_arg(item.get("name")).strip(): _coerce_text_arg(item.get("type")).strip().lower()
-        for item in dataset_schema
-        if isinstance(item, dict) and _coerce_text_arg(item.get("name")).strip()
-    }
-
-
-def _is_numeric_schema_type(type_name: str) -> bool:
-    normalized = str(type_name or "").strip().lower()
-    return any(
-        token in normalized
-        for token in ("int", "float", "double", "decimal", "numeric", "real", "bigint", "smallint")
-    )
-
-
-def _is_datetime_schema_type(dtype: str) -> bool:
-    normalized = dtype.lower()
-    return "date" in normalized or "time" in normalized
-
-
-def _build_dashboard_chart_specs(
-    normalized_bindings: Dict[int, Dict[str, Any]],
-    dataset_schema: List[Dict[str, Any]],
-) -> Dict[int, Dict[str, Any]]:
-    schema_types = _schema_type_map(dataset_schema)
-    normalized_specs: Dict[int, Dict[str, Any]] = {}
-    for chart_index, binding in normalized_bindings.items():
-        x_field = _coerce_text_arg(binding.get("x_field")).strip()
-        y_field = _coerce_text_arg(binding.get("y_field")).strip()
-        dimension_field = _coerce_text_arg(binding.get("dimension_field")).strip() or x_field
-        metric_field = _coerce_text_arg(binding.get("metric_field")).strip() or y_field
-        orientation = _coerce_text_arg(binding.get("orientation")).strip().lower()
-        numerator_field = _coerce_text_arg(binding.get("numerator_field")).strip()
-        denominator_field = _coerce_text_arg(binding.get("denominator_field")).strip()
-
-        if orientation == "h" and x_field and y_field:
-            x_is_numeric = _is_numeric_schema_type(schema_types.get(x_field, ""))
-            y_is_numeric = _is_numeric_schema_type(schema_types.get(y_field, ""))
-            if x_is_numeric and not y_is_numeric:
-                dimension_field = y_field
-                metric_field = x_field
-
-        aggregation = _coerce_text_arg(binding.get("aggregation"), "count").strip().lower() or "count"
-        if numerator_field and denominator_field and aggregation in {"avg", "mean"}:
-            aggregation = "ratio"
-
-        live_capable = bool(
-            dimension_field
-            and (
-                aggregation == "count"
-                or metric_field
-                or (numerator_field and denominator_field)
-            )
-        )
-        chart_id = _coerce_text_arg(binding.get("chart_id"), f"chart_{chart_index}").strip() or f"chart_{chart_index}"
-        normalized_specs[chart_index] = {
-            "chartId": chart_id,
-            "chartIndex": chart_index,
-            "chartType": _coerce_text_arg(binding.get("chart_type"), "bar").strip().lower() or "bar",
-            "dimensionField": dimension_field,
-            "metricField": metric_field,
-            "numeratorField": numerator_field,
-            "denominatorField": denominator_field,
-            "aggregation": aggregation,
-            "orientation": orientation,
-            "seriesField": _coerce_text_arg(binding.get("series_field")).strip(),
-            "sort": {
-                "by": _coerce_text_arg(binding.get("sort_by"), "y").strip().lower() or "y",
-                "direction": _coerce_text_arg(binding.get("sort_direction"), "desc").strip().lower() or "desc",
-            },
-            "limit": _coerce_int_arg(binding.get("limit"), 0, minimum=0),
-            "mode": _coerce_text_arg(binding.get("mode"), "lines+markers").strip(),
-            "labels": {
-                "title": _coerce_text_arg(binding.get("title")).strip(),
-                "x": _coerce_text_arg(binding.get("x_title")).strip(),
-                "y": _coerce_text_arg(binding.get("y_title")).strip(),
-            },
-            "questionAnswered": _coerce_text_arg(binding.get("question_answered")).strip(),
-            "whyItMatters": _coerce_text_arg(binding.get("why_it_matters")).strip(),
-            "highlightRule": _coerce_text_arg(binding.get("highlight_rule")).strip(),
-            "format": _coerce_text_arg(binding.get("format")).strip(),
-            "interactive": _coerce_bool_arg(binding.get("interactive"), True),
-            "layoutSpan": _coerce_text_arg(binding.get("layout_span"), "half").strip().lower() or "half",
-            "layoutSection": _coerce_text_arg(binding.get("layout_section"), "drivers").strip().lower() or "drivers",
-            "filters": list(binding.get("filters") or []),
-            "liveCapable": live_capable,
-            "snapshotOnly": not live_capable,
-            "staticReason": _coerce_text_arg(binding.get("static_reason")).strip(),
-        }
-    return normalized_specs
-
-
-def _validate_dashboard_runtime_config(
-    *,
-    filter_config: List[Dict[str, Any]],
-    chart_bindings: Dict[int, Dict[str, Any]],
-    dataset_schema: List[Dict[str, Any]],
-) -> List[str]:
-    errors: List[str] = []
-    fields = _schema_type_map(dataset_schema)
-    field_names = set(fields)
-
-    for filter_def in filter_config:
-        field = _coerce_text_arg(filter_def.get("field")).strip()
-        if field and field not in field_names:
-            errors.append(f"Filter field '{field}' is not present in the dashboard dataset.")
-
-    for chart_index, binding in sorted(chart_bindings.items()):
-        prefix = f"Chart {chart_index}"
-        chart_type = _coerce_text_arg(binding.get("chart_type")).strip().lower()
-        aggregation = _coerce_text_arg(binding.get("aggregation")).strip().lower()
-        x_field = _coerce_text_arg(binding.get("x_field")).strip()
-        y_field = _coerce_text_arg(binding.get("y_field")).strip()
-        dimension_field = _coerce_text_arg(binding.get("dimension_field")).strip() or x_field
-        metric_field = _coerce_text_arg(binding.get("metric_field")).strip() or y_field
-        numerator_field = _coerce_text_arg(binding.get("numerator_field")).strip()
-        denominator_field = _coerce_text_arg(binding.get("denominator_field")).strip()
-        series_field = _coerce_text_arg(binding.get("series_field")).strip()
-        sort_by = _coerce_text_arg(binding.get("sort_by")).strip().lower()
-        sort_direction = _coerce_text_arg(binding.get("sort_direction")).strip().lower()
-        orientation = _coerce_text_arg(binding.get("orientation")).strip().lower()
-
-        if chart_type not in NATIVE_DASHBOARD_CHART_TYPES:
-            errors.append(
-                f"{prefix} uses unsupported chart_type '{chart_type}'. "
-                f"Use one of: {', '.join(sorted(NATIVE_DASHBOARD_CHART_TYPES))}."
-            )
-        if aggregation not in NATIVE_DASHBOARD_AGGREGATIONS and not (numerator_field and denominator_field):
-            errors.append(
-                f"{prefix} uses unsupported aggregation '{aggregation}'. "
-                f"Use one of: {', '.join(sorted(NATIVE_DASHBOARD_AGGREGATIONS))}."
-            )
-        if sort_by not in NATIVE_DASHBOARD_SORT_FIELDS:
-            errors.append(f"{prefix} sort_by must be 'x' or 'y'.")
-        if sort_direction not in NATIVE_DASHBOARD_SORT_DIRECTIONS:
-            errors.append(f"{prefix} sort_direction must be 'asc' or 'desc'.")
-        if orientation not in NATIVE_DASHBOARD_ORIENTATIONS:
-            errors.append(f"{prefix} orientation must be '', 'h', or 'v'.")
-        if not dimension_field:
-            errors.append(f"{prefix} is missing x_field or dimension_field.")
-        elif dimension_field not in field_names:
-            errors.append(f"{prefix} dimension field '{dimension_field}' is not present in the dashboard dataset.")
-        if aggregation != "count" and not numerator_field and not denominator_field:
-            if not metric_field:
-                errors.append(f"{prefix} is missing y_field or metric_field for aggregation '{aggregation}'.")
-            elif metric_field not in field_names:
-                errors.append(f"{prefix} metric field '{metric_field}' is not present in the dashboard dataset.")
-        for role, field in (
-            ("numerator_field", numerator_field),
-            ("denominator_field", denominator_field),
-            ("series_field", series_field),
-        ):
-            if field and field not in field_names:
-                errors.append(f"{prefix} {role} '{field}' is not present in the dashboard dataset.")
-        if series_field and series_field == dimension_field:
-            errors.append(f"{prefix} uses the same field for dimension and series; choose a different grouping.")
-        if chart_type in {"line", "area", "scatter"} and dimension_field and _is_datetime_schema_type(fields.get(dimension_field, "")):
-            if sort_by != "x" or sort_direction != "asc":
-                errors.append(f"{prefix} is time-based and must use sort_by='x' with sort_direction='asc'.")
-
-    return errors
-
-
-_GENERIC_DASHBOARD_TITLE_TOKENS = {
-    "chart",
-    "overview",
-    "analysis",
-    "trends",
-    "breakdown",
-    "geographic trends",
-    "top categories",
-    "browser/device segmentation",
-    "country comparison",
-}
-
-
-def _looks_schema_like_title(value: str) -> bool:
-    normalized = re.sub(r"\s+", " ", _coerce_text_arg(value).strip().lower())
-    if not normalized:
-        return True
-    if normalized in _GENERIC_DASHBOARD_TITLE_TOKENS:
-        return True
-    return bool(re.fullmatch(r"[a-z0-9_ /-]+", normalized) and "_" in normalized)
-
-
-def _validate_dashboard_spec_inputs(
-    *,
-    strict_dashboard_mode: bool,
-    dashboard_dataset_path: str,
-    audience: str,
-    business_question: str,
-    decision_questions: List[str],
-    layout_template: str,
-    metric_cards: List[Dict[str, str]],
-    chart_specs: Dict[int, Dict[str, Any]],
-) -> Optional[str]:
-    if not strict_dashboard_mode:
-        return None
-    if not chart_specs:
-        return (
-            "Dashboard package incomplete: no structured charts were provided. "
-            "Pass approved chart bindings so the live runtime has charts to render."
-        )
-    if not _coerce_text_arg(dashboard_dataset_path).strip():
-        return (
-            "Dashboard package incomplete: dashboard_dataset_path is required so the live runtime "
-            "and snapshot share the same canonical dataset."
-        )
-    if not audience.strip():
-        return "Dashboard spec incomplete: audience is required for executive dashboard mode."
-    if not business_question.strip():
-        return "Dashboard spec incomplete: business_question is required for executive dashboard mode."
-    if not decision_questions:
-        return "Dashboard spec incomplete: provide at least one decision question for executive dashboard mode."
-    if not layout_template.strip():
-        return "Dashboard spec incomplete: layout_template is required for executive dashboard mode."
-    if not metric_cards or len(metric_cards) > 3:
-        return "Dashboard spec incomplete: provide 2 to 3 KPI cards for the executive hero."
-    if len(chart_specs) < STRICT_DASHBOARD_MIN_CHART_COUNT or len(chart_specs) > MAX_CHART_COUNT:
-        return (
-            f"Dashboard spec incomplete: provide {STRICT_DASHBOARD_MIN_CHART_COUNT} to {MAX_CHART_COUNT} "
-            "structured charts for executive dashboard mode."
-        )
-    for chart_spec in chart_specs.values():
-        title = _coerce_text_arg((chart_spec.get("labels") or {}).get("title"))
-        if _looks_schema_like_title(title):
-            return (
-                "Dashboard spec validation failed: chart titles must be business-readable and insight-led, "
-                f"not generic placeholders ('{title or 'untitled'}')."
-            )
-        if not _coerce_text_arg(chart_spec.get("questionAnswered")).strip():
-            return f"Dashboard spec validation failed: {title} is missing questionAnswered."
-        if not _coerce_text_arg(chart_spec.get("whyItMatters")).strip():
-            return f"Dashboard spec validation failed: {title} is missing whyItMatters."
-    return None
-
-
-def _emit_dashboard_artifact(
-    callbacks: Optional[CallbackManagerForToolRun],
-    payload: Dict[str, Any],
-) -> None:
-    if not callbacks or not payload:
-        return
-    try:
-        run_id = getattr(callbacks, "run_id", None)
-        if run_id is not None:
-            callbacks.on_custom_event("dashboard_artifact", payload, run_id=run_id)
-        else:
-            callbacks.on_custom_event("dashboard_artifact", payload)
-    except Exception:  # pragma: no cover - best effort
-        logger.warning("Failed to dispatch dashboard_artifact event", exc_info=True)
-
-
-def _looks_like_local_file_query(sql_query: str) -> bool:
-    normalized = (sql_query or "").strip().lower()
-    if not normalized:
-        return False
-    if "read_parquet" in normalized or "read_csv_auto" in normalized:
-        return True
-    return bool(re.search(r"['\"]/?[^'\"]+\.(parquet|csv|jsonl?|ndjson)['\"]", sql_query, re.IGNORECASE))
-
-
-def _cache_key_for_query(sql_query: str, workspace_id: str, cache_key_hint: str) -> str:
-    digest = hashlib.sha256(
-        f"{workspace_id}\n{cache_key_hint}\n{sql_query}".encode("utf-8")
-    ).hexdigest()
-    return digest[:12]
-
-
-def _schema_summary_from_dataframe(df: pd.DataFrame) -> List[Dict[str, Any]]:
-    return [
-        {
-            "name": str(column),
-            "type": str(dtype),
-            "mode": "NULLABLE",
-        }
-        for column, dtype in df.dtypes.items()
-    ]
-
-
-def _load_dataframe_from_parquet(connection: duckdb.DuckDBPyConnection, parquet_path: Path) -> pd.DataFrame:
-    safe_path = parquet_path.as_posix().replace("'", "''")
-    return connection.execute(f"SELECT * FROM read_parquet('{safe_path}')").df()
-
-
-def _emit_artifacts(
-    callbacks: Optional[CallbackManagerForToolRun],
-    artifacts: List[Dict[str, Any]],
-) -> None:
-    if not callbacks or not artifacts:
-        return
-    try:
-        run_id = getattr(callbacks, "run_id", None)
-        payload = {"files": artifacts}
-        if run_id is not None:
-            callbacks.on_custom_event("tool_artifacts", payload, run_id=run_id)
-        else:
-            callbacks.on_custom_event("tool_artifacts", payload)
-    except Exception:  # pragma: no cover - best effort
-        logger.warning("Failed to dispatch tool_artifacts event", exc_info=True)
-
-
-class DuckDBManager:
-    """Manages DuckDB connection and file registration."""
-
-    def __init__(self, workspace_state: WorkspaceState):
-        self.workspace_state = workspace_state
-        self.con = duckdb.connect(database=":memory:")
-        # The agent can emit multiple local SQL tool calls in one model step. DuckDB's
-        # Python connection is not safe for overlapping use, so serialize all access
-        # through a per-manager lock to avoid orphaned/stuck runs.
-        self._con_lock = threading.RLock()
-        self.session = DataAgentSessionState()
-        self._registered_tables: Set[str] = set()
-        self._register_files()
-
-    def reset_session(self):
-        """Reset per-run guardrails."""
-        self.session.reset()
-        self._register_files()
-
-    def _register_files(self):
-        """Scans workspace for CSV and Parquet files and registers them as tables."""
-        with self._con_lock:
-            self._registered_tables.clear()
-            root = self.workspace_state.root_path
-            preferred_dirs = tuple(
-                dirname
-                for dirname in DATA_DISCOVERY_DIR_CANDIDATES
-                if (root / dirname).exists()
-            )
-            data_files = _iter_workspace_files(
-                root,
-                allowed_extensions=DATA_FILE_EXTENSIONS,
-                preferred_dirs=preferred_dirs,
-            )
-            used_names: Set[str] = set()
-            for file_path in sorted(data_files):
-                relative_stem = file_path.relative_to(root).with_suffix("").as_posix()
-                base_name = re.sub(r"[^a-zA-Z0-9_]", "_", file_path.stem)
-                path_name = re.sub(r"[^a-zA-Z0-9_]", "_", relative_stem)
-                ext_name = f"{base_name}_{file_path.suffix.lstrip('.').lower()}"
-                table_name = base_name
-                if table_name in used_names:
-                    table_name = path_name or ext_name
-                if table_name in used_names:
-                    table_name = ext_name
-                safe_path = file_path.as_posix().replace("'", "''")
-                try:
-                    if file_path.suffix.lower() == ".csv":
-                        self.con.execute(
-                            f"CREATE OR REPLACE TABLE {table_name} "
-                            f"AS SELECT * FROM read_csv_auto('{safe_path}')"
-                        )
-                    else:
-                        self.con.execute(
-                            f"CREATE OR REPLACE TABLE {table_name} "
-                            f"AS SELECT * FROM read_parquet('{safe_path}')"
-                        )
-                    used_names.add(table_name)
-                    self._registered_tables.add(table_name)
-                    logger.info("Registered table %s from %s", table_name, file_path)
-                except Exception as exc:
-                    logger.error("Failed to register %s: %s", file_path, exc)
-
-    def get_schema(self, table_names: Optional[List[str]] = None) -> str:
-        if _is_strict_dashboard_mode(self.workspace_state):
-            if self.session.last_schema_result is not None and self.session.schema_read_count >= STRICT_DASHBOARD_SCHEMA_COUNT:
-                return (
-                    "Dashboard planning mode reuses the existing schema snapshot to avoid duplicated discovery.\n\n"
-                    f"{self.session.last_schema_result}"
-                )
-        with self._con_lock:
-            tables = self.con.execute("SHOW TABLES").fetchall()
-            if not tables:
-                return "No tables found in the workspace."
-
-            schema_lines: List[str] = []
-            for table in tables:
-                table_name = table[0]
-                if table_names and table_name not in table_names:
-                    continue
-
-                schema_lines.append(f"Table: {table_name}")
-                columns = self.con.execute(f"DESCRIBE {table_name}").fetchall()
-                sample_rows = self.con.execute(f"SELECT * FROM {table_name} LIMIT 3").df()
-                for col in columns:
-                    sample_values: List[str] = []
-                    if not sample_rows.empty and col[0] in sample_rows.columns:
-                        non_null = sample_rows[col[0]].dropna().tolist()
-                        for value in non_null:
-                            formatted = _format_sample_value(value)
-                            if formatted not in sample_values:
-                                sample_values.append(formatted)
-                            if len(sample_values) == 2:
-                                break
-                    sample_suffix = (
-                        f" [examples: {', '.join(sample_values)}]" if sample_values else ""
-                    )
-                    schema_lines.append(f"  - {col[0]} ({col[1]}){sample_suffix}")
-                schema_lines.append("")
-
-            rendered = "\n".join(schema_lines).strip()
-            self.session.schema_read_count += 1
-            self.session.last_schema_result = rendered
-            return rendered
-
-    def run_query(
-        self,
-        query: str,
-        *,
-        record_sql: Optional[str] = None,
-        truncated: bool = False,
-    ) -> pd.DataFrame:
-        with self._con_lock:
-            strict_dashboard_mode = _is_strict_dashboard_mode(self.workspace_state)
-            max_query_count = STRICT_DASHBOARD_QUERY_COUNT if strict_dashboard_mode else MAX_QUERY_COUNT
-            if self.session.query_count >= max_query_count:
-                raise ValueError(
-                    f"Query budget exhausted: at most {max_query_count} queries are allowed per run."
-                )
-            df = self.con.execute(query).df()
-            self.session.query_count += 1
-            self.session.last_query_result = df
-            stored_sql = record_sql or query
-            self.session.last_query_sql = stored_sql
-            if strict_dashboard_mode:
-                signature = _extract_dashboard_dimension_signature(stored_sql)
-                if signature:
-                    self.session.dashboard_dimension_signatures.add(signature)
-            self.session.query_history.append(
-                _QueryRecord(
-                    sql=stored_sql,
-                    row_count=len(df),
-                    preview=df.head(MAX_RESULT_ROWS).copy(),
-                    truncated=truncated,
-                )
-            )
-            return df
-
-    def record_chart(self, title: str, artifact_paths: List[str]) -> None:
-        self.session.chart_count += 1
-        self.session.chart_history.append(_ChartRecord(title=title, artifact_paths=artifact_paths))
-
-    def record_materialization(self, record: _MaterializationRecord) -> None:
-        self.session.materialization_history.append(record)
-
-    def register_artifact(self, artifact: Dict[str, Any]) -> None:
-        self.session.run_artifacts.append(artifact)
-
-    def refresh_registered_files(self) -> None:
-        self._register_files()
-
-    def require_schema_check(self) -> None:
-        if not self.session.schema_inspected:
-            raise ValueError(
-                "Call get_table_schema before running SQL queries to verify the data layout."
-            )
-
-    def require_query_before_chart(self) -> None:
-        if self.session.last_query_result is None:
-            raise ValueError("Run at least one SQL query before generating a chart.")
-
-    def require_chart_budget(self) -> None:
-        if self.session.chart_count >= MAX_CHART_COUNT:
-            raise ValueError(
-                f"Chart budget exhausted: at most {MAX_CHART_COUNT} charts are allowed per run."
-            )
-
-    def ensure_single_summary(self) -> None:
-        if self.session.dashboard_generated:
-            raise ValueError("A dashboard has already been generated for this run.")
-        if self.session.summary_generated:
-            raise ValueError("A summary has already been generated for this run.")
-
-    def ensure_single_dashboard(self) -> None:
-        if self.session.summary_generated:
-            raise ValueError("A summary has already been generated for this run.")
-        if self.session.dashboard_generated:
-            raise ValueError("A dashboard has already been generated for this run.")
-
-    def mark_summary_generated(self) -> None:
-        self.session.summary_generated = True
-
-    def mark_dashboard_generated(self) -> None:
-        self.session.dashboard_generated = True
-
-    def get_limited_result(self) -> Optional[pd.DataFrame]:
-        result = self.session.last_query_result
-        if result is None:
-            return None
-        if len(result) > MAX_SESSION_ROWS:
-            return result.head(MAX_SESSION_ROWS).copy()
-        return result.copy()
-
-
-_STRING_LITERAL_PATTERN = re.compile(r"(['\"])([^'\"\\]*)\1")
-
-
-def _rewrite_virtual_paths(query: str, workspace_root: Path) -> str:
-    """Rewrite absolute paths that assume the workspace root is '/' to real filesystem paths."""
-
-    def replacer(match: re.Match[str]) -> str:
-        quote, content = match.group(1), match.group(2)
-        if content.startswith("/"):
-            resolved = (workspace_root / content.lstrip("/")).resolve()
-            return f"{quote}{resolved}{quote}"
-        return match.group(0)
-
-    return _STRING_LITERAL_PATTERN.sub(replacer, query)
-
-
-def sanitize_python_code(code: str) -> str:
-    """Sanitize Python code to prevent malicious execution."""
-    code = re.sub(r"#.*$", "", code, flags=re.MULTILINE)
-    code = re.sub(r'"""[\s\S]*?"""', "", code)
-    code = re.sub(r"'''[\s\S]*?'''", "", code)
-
-    forbidden_patterns = [
-        r"import\s+os", r"import\s+sys", r"import\s+subprocess", r"import\s+__builtins__",
-        r"eval\s*\(", r"exec\s*\(", r"open\s*\(", r"file\s*\(",
-        r"input\s*\(", r"raw_input\s*\(", r"compile\s*\(", r"globals\s*\(",
-        r"locals\s*\(", r"vars\s*\(", r"dir\s*\(", r"help\s*\(", r"breakpoint\s*\(",
-        r"quit\s*\(", r"exit\s*\(", r"while\s+True:",
-        r"async\s+", r"await\s+",
-        r"pd\.read_csv", r"pd\.read_parquet", r"pd\.read_json", r"pd\.read_excel",
-        r"read_csv_auto", r"read_parquet",
-    ]
-
-    for pattern in forbidden_patterns:
-        if re.search(pattern, code, re.IGNORECASE):
-            raise ValueError(f"Forbidden pattern detected: {pattern}")
-
-    allowed_imports = [
-        "pandas",
-        "pd",
-        "numpy",
-        "np",
-        "json",
-        "math",
-        "statistics",
-        "plotly",
-        "plotly.express",
-        "plotly.graph_objects",
-        "plotly.io",
-        "altair",
-        "altair.vegalite.v5",
-    ]
-
-    try:
-        tree = ast.parse(code)
-        for node in ast.walk(tree):
-            if isinstance(node, ast.Import):
-                for alias in node.names:
-                    if alias.name not in allowed_imports:
-                        raise ValueError(f"Forbidden import: {alias.name}")
-            elif isinstance(node, ast.ImportFrom):
-                if node.module not in allowed_imports:
-                    raise ValueError(f"Forbidden import from: {node.module}")
-    except SyntaxError as exc:
-        raise ValueError(f"Invalid Python syntax: {exc}")
-
-    return code
-
-
-def build_data_agent_tools(workspace_state: WorkspaceState, source_tracker: Any = None) -> List[Tool]:
-    db_manager = DuckDBManager(workspace_state)
-    workspace_state.context["data_agent_manager"] = db_manager
-
-    @tool
-    def get_table_schema(
-        table_names: List[str] = Field(description="List of table names to get schema for. If empty, returns all tables."),
-    ) -> str:
-        """Get the schema of tables relevant to the user question."""
-        result = db_manager.get_schema(table_names if table_names else None)
-        db_manager.session.schema_inspected = True
-        return result
-
-    @tool
-    def run_sql_query(
-        sql_query: str = Field(description="The SQL query to run."),
-    ) -> str:
-        """Run a SQL query against the database and return the results."""
-        try:
-            db_manager.require_schema_check()
-        except ValueError as exc:
-            return str(exc)
-
-        try:
-            cleaned_query = sql_query.strip()
-            if cleaned_query.endswith(";"):
-                cleaned_query = cleaned_query[:-1].rstrip()
-            strict_dashboard_mode = _is_strict_dashboard_mode(workspace_state)
-            if strict_dashboard_mode and not _is_plan_approved(workspace_state):
-                if db_manager.session.query_count >= STRICT_DASHBOARD_PREVIEW_QUERY_COUNT:
-                    return (
-                        f"{_dashboard_plan_gate_message()} Only {STRICT_DASHBOARD_PREVIEW_QUERY_COUNT} preview query "
-                        "is allowed before approval."
-                    )
-                if not _looks_like_preview_query(cleaned_query):
-                    return (
-                        f"{_dashboard_plan_gate_message()} Before approval, only one lightweight preview query "
-                        "with a small LIMIT is allowed."
-                    )
-            if strict_dashboard_mode and _is_plan_approved(workspace_state):
-                signature = _extract_dashboard_dimension_signature(cleaned_query)
-                if signature and signature in db_manager.session.dashboard_dimension_signatures:
-                    return (
-                        "Duplicate dashboard dimension pass blocked. Reuse the existing aggregate for this dimension "
-                        "or update the approved plan before querying it again."
-                    )
-            cleaned_query = _rewrite_virtual_paths(
-                cleaned_query, db_manager.workspace_state.root_path
-            )
-            truncated = False
-            executed_query = cleaned_query
-            safety_capped = not _query_looks_aggregated(cleaned_query)
-            if safety_capped:
-                executed_query = (
-                    f"SELECT * FROM ({cleaned_query}) AS helpudoc_query_preview "
-                    f"LIMIT {MAX_QUERY_RESULT_ROWS}"
-                )
-
-            df = db_manager.run_query(executed_query, record_sql=cleaned_query)
-            if safety_capped and len(df) > MAX_SESSION_ROWS:
-                truncated = True
-                df = df.head(MAX_SESSION_ROWS).copy()
-                db_manager.session.last_query_result = df
-                if db_manager.session.query_history:
-                    db_manager.session.query_history[-1].row_count = len(df)
-                    db_manager.session.query_history[-1].preview = df.head(MAX_RESULT_ROWS).copy()
-                    db_manager.session.query_history[-1].truncated = True
-            return _format_dataframe_markdown(df, truncated=truncated)
-        except Exception as exc:  # pragma: no cover - defensive
-            return f"Error executing query: {str(exc)}"
-
-    @tool
-    def materialize_bigquery_to_parquet(
-        sql_query: Annotated[
-            str,
-            Field(description="Read-only BigQuery SQL query to materialize into workspace-local Parquet."),
-        ],
-        cache_key_hint: Annotated[
-            str,
-            Field(description="Optional stable label to make cache files easier to recognize."),
-        ] = "",
-        ttl_hours: Annotated[
-            int,
-            Field(description="Hours before the cached Parquet is considered stale and should be refreshed."),
-        ] = DEFAULT_CACHE_TTL_HOURS,
-        force_refresh: Annotated[
-            bool,
-            Field(description="When true, ignore any cached Parquet and re-run the warehouse query."),
-        ] = False,
-        max_rows: Annotated[
-            int,
-            Field(description="Safety cap for exported rows. Keep results scoped for iterative analysis."),
-        ] = MAX_MATERIALIZED_ROWS,
-        target_path: Annotated[
-            str,
-            Field(description="Optional stable parquet path such as datasets/orders/latest.parquet."),
-        ] = "",
-        emit_csv: Annotated[
-            bool,
-            Field(description="When true, also publish a CSV mirror of the refreshed dataset."),
-        ] = False,
-        csv_path: Annotated[
-            str,
-            Field(description="Optional stable CSV path such as datasets/orders/latest.csv."),
-        ] = "",
-        callbacks: Optional[CallbackManagerForToolRun] = None,
-    ) -> str:
-        """Materialize a read-only BigQuery query into cached and optionally stable workspace Parquet."""
-        blocked = tagged_files_mode_guard(workspace_state.context, "cache_bigquery_query")
-        if blocked:
-            return blocked
-        if workspace_state.context.get("tagged_files_only"):
-            return "Tool disabled: tagged files were provided, use rag_query only."
-        if _is_strict_dashboard_mode(workspace_state):
-            return (
-                "Dashboard planning mode is bound to tagged local datasets. "
-                "Skip warehouse rediscovery and use the tagged parquet/csv as the source of truth."
-            )
-
-        normalized_sql = _coerce_text_arg(sql_query).strip().rstrip(";")
-        if not normalized_sql:
-            return "SQL query is required."
-        if _looks_like_local_file_query(normalized_sql):
-            return (
-                "materialize_bigquery_to_parquet expects BigQuery SQL, not workspace file paths. "
-                "Use get_table_schema and run_sql_query against the DuckDB-registered local table instead."
-            )
-
-        try:
-            validate_read_only_sql(normalized_sql)
-        except ValueError as exc:
-            return str(exc)
-
-        hint = _coerce_text_arg(cache_key_hint).strip()
-        ttl_value = _coerce_int_arg(ttl_hours, DEFAULT_CACHE_TTL_HOURS, minimum=1)
-        row_cap = _coerce_int_arg(
-            max_rows,
-            MAX_MATERIALIZED_ROWS,
-            minimum=1,
-            maximum=MAX_MATERIALIZED_ROWS,
-        )
-        refresh = _coerce_bool_arg(force_refresh, False)
-        publish_csv = _coerce_bool_arg(emit_csv, False)
-        toolbox_defaults = load_bigquery_toolbox_config()
-        preferred_server = str(toolbox_defaults.get("server_name") or "toolbox-bq-demo")
-        project = _coerce_text_arg(
-            workspace_state.context.get("bigquery_project")
-            or workspace_state.context.get("bq_project")
-            or toolbox_defaults.get("project")
-        ).strip()
-        location = _coerce_text_arg(
-            workspace_state.context.get("bigquery_location")
-            or workspace_state.context.get("bq_location")
-            or toolbox_defaults.get("location")
-        ).strip()
-
-        auth_header = extract_bearer_header(workspace_state, preferred_server)
-        if not auth_header:
-            return (
-                "BigQuery materialization is unavailable because no delegated BigQuery access token was found. "
-                f"Expected MCP auth for server '{preferred_server}' or BQ_ACCESS_TOKEN."
-            )
-
-        cache_key = _cache_key_for_query(
-            normalized_sql,
-            _coerce_text_arg(getattr(workspace_state, "workspace_id", "")),
-            hint,
-        )
-        cache_slug = _safe_slug(hint or "bq_export", "bq_export")
-        cache_dir = workspace_state.root_path / "data_cache" / "bigquery"
-        cache_dir.mkdir(parents=True, exist_ok=True)
-        cache_parquet_path = cache_dir / f"{cache_slug}_{cache_key}.parquet"
-        cache_metadata_path = cache_dir / f"{cache_slug}_{cache_key}.metadata.json"
-
-        now = _utc_now()
-        cached = False
-        metadata_payload: Dict[str, Any] = {}
-        df: Optional[pd.DataFrame] = None
-
-        if not refresh and cache_parquet_path.exists() and cache_metadata_path.exists():
-            try:
-                metadata_payload = json.loads(cache_metadata_path.read_text(encoding="utf-8"))
-                expires_at_raw = _coerce_text_arg(metadata_payload.get("expiresAt")).strip()
-                expires_at = datetime.fromisoformat(expires_at_raw.replace("Z", "+00:00")) if expires_at_raw else None
-                if expires_at and expires_at > now:
-                    cached = True
-            except Exception:
-                cached = False
-                metadata_payload = {}
-
-        if not cached:
-            try:
-                df = run_bigquery_query(
-                    sql=normalized_sql,
-                    project=project,
-                    location=location,
-                    auth_header=auth_header,
-                    row_limit=row_cap,
-                )
-                write_export_dataframe(df, cache_parquet_path, "parquet")
-            except Exception as exc:  # pragma: no cover - network/filesystem guard
-                logger.exception("BigQuery materialization failed")
-                return f"BigQuery materialization failed: {exc}"
-
-            expires_at = now + timedelta(hours=ttl_value)
-            metadata_payload = {
-                "cacheKey": cache_key,
-                "cacheKeyHint": hint,
-                "connector": preferred_server,
-                "project": project,
-                "location": location,
-                "sourceSql": normalized_sql,
-                "rowCount": len(df),
-                "schema": _schema_summary_from_dataframe(df),
-                "refreshedAt": now.isoformat().replace("+00:00", "Z"),
-                "materializedAt": now.isoformat().replace("+00:00", "Z"),
-                "expiresAt": expires_at.isoformat().replace("+00:00", "Z"),
-                "ttlHours": ttl_value,
-                "parquetPath": _workspace_rel(cache_parquet_path, workspace_state.root_path),
-                "csvPath": "",
-                "artifactTargets": {"dashboard": "", "report": ""},
-            }
-            cache_metadata_path.write_text(_json_dump(metadata_payload), encoding="utf-8")
-
-        stable_parquet_path: Optional[Path] = None
-        stable_csv_path: Optional[Path] = None
-        manifest_path = cache_metadata_path
-        artifact_paths: List[Path] = [cache_parquet_path, cache_metadata_path]
-
-        if target_path.strip():
-            try:
-                stable_parquet_path = resolve_output_path(
-                    workspace_state.root_path,
-                    target_path,
-                    "parquet",
-                )
-            except ValueError as exc:
-                return str(exc)
-            stable_parquet_path.parent.mkdir(parents=True, exist_ok=True)
-            shutil.copy2(cache_parquet_path, stable_parquet_path)
-            manifest_path = stable_parquet_path.parent / "manifest.json"
-            artifact_paths.append(stable_parquet_path)
-
-        if publish_csv:
-            if df is None:
-                df = _load_dataframe_from_parquet(db_manager.con, cache_parquet_path)
-            raw_csv_path = csv_path.strip()
-            if not raw_csv_path and stable_parquet_path is not None:
-                raw_csv_path = (stable_parquet_path.parent / "latest.csv").relative_to(
-                    workspace_state.root_path
-                ).as_posix()
-            try:
-                stable_csv_path = resolve_output_path(
-                    workspace_state.root_path,
-                    raw_csv_path,
-                    "csv",
-                )
-            except ValueError as exc:
-                return str(exc)
-            write_export_dataframe(df, stable_csv_path, "csv")
-            artifact_paths.append(stable_csv_path)
-
-        if stable_parquet_path is not None:
-            stable_metadata = dict(metadata_payload)
-            stable_metadata["parquetPath"] = _workspace_rel(stable_parquet_path, workspace_state.root_path)
-            stable_metadata["csvPath"] = (
-                _workspace_rel(stable_csv_path, workspace_state.root_path)
-                if stable_csv_path is not None
-                else ""
-            )
-            stable_metadata["artifactTargets"] = {
-                "dashboard": "",
-                "report": "",
-            }
-            manifest_path.write_text(_json_dump(stable_metadata), encoding="utf-8")
-            artifact_paths.append(manifest_path)
-        else:
-            if stable_csv_path is not None:
-                metadata_payload["csvPath"] = _workspace_rel(stable_csv_path, workspace_state.root_path)
-                cache_metadata_path.write_text(_json_dump(metadata_payload), encoding="utf-8")
-
-        db_manager.refresh_registered_files()
-
-        registered_artifacts: List[Dict[str, Any]] = []
-        for artifact_path in artifact_paths:
-            if not artifact_path.exists():
-                continue
-            mime_type = ALLOWED_ARTIFACT_EXTENSIONS.get(artifact_path.suffix.lower())
-            if not mime_type:
-                continue
-            artifact_payload = {
-                "path": _workspace_rel(artifact_path, workspace_state.root_path),
-                "mimeType": mime_type,
-                "size": artifact_path.stat().st_size,
-            }
-            db_manager.register_artifact(artifact_payload)
-            registered_artifacts.append(artifact_payload)
-
-        published_parquet = stable_parquet_path or cache_parquet_path
-        published_metadata = manifest_path
-        db_manager.record_materialization(
-            _MaterializationRecord(
-                cache_key=cache_key,
-                sql=normalized_sql,
-                parquet_path=_workspace_rel(published_parquet, workspace_state.root_path),
-                metadata_path=_workspace_rel(published_metadata, workspace_state.root_path),
-                row_count=_coerce_int_arg(metadata_payload.get("rowCount"), 0, minimum=0),
-                connector=preferred_server,
-                cached=cached,
-                expires_at=_coerce_text_arg(metadata_payload.get("expiresAt")),
-            )
-        )
-        _emit_artifacts(callbacks, registered_artifacts)
-
-        payload = {
-            "cache_key": cache_key,
-            "cached": cached,
-            "connector": preferred_server,
-            "row_count": _coerce_int_arg(metadata_payload.get("rowCount"), 0, minimum=0),
-            "schema": list(metadata_payload.get("schema") or []),
-            "parquet_path": _workspace_rel(published_parquet, workspace_state.root_path),
-            "metadata_path": _workspace_rel(published_metadata, workspace_state.root_path),
-            "csv_path": (
-                _workspace_rel(stable_csv_path, workspace_state.root_path)
-                if stable_csv_path is not None
-                else _coerce_text_arg(metadata_payload.get("csvPath"))
-            ),
-            "duckdb_table_name": published_parquet.stem,
-            "expires_at": metadata_payload.get("expiresAt"),
-            "ttl_hours": ttl_value,
-            "cache_parquet_path": _workspace_rel(cache_parquet_path, workspace_state.root_path),
-            "cache_metadata_path": _workspace_rel(cache_metadata_path, workspace_state.root_path),
-        }
-        return _json_dump(payload)
-
-    @tool
-    def generate_chart_config(
-        chart_title: str = Field(description="The title of the chart"),
-        python_code: str = Field(
-            description=(
-                "Python code to create visualizations from the provided df variable, which contains "
-                "the last run_sql_query result. Do not read files directly; use DuckDB via "
-                "run_sql_query before calling this tool. Two approaches:\n"
-                "1. Matplotlib/Seaborn: Use plt.figure(), plt.plot(), sns.barplot(), etc. "
-                "Figures are auto-saved as PNG.\n"
-                "2. Plotly (preferred): Build a Plotly figure or a dict with data/layout/config assigned to 'chart_config'. "
-                "Plotly specs are saved as .plotly.json for the viewer."
-            )
-        ),
-        callbacks: Optional[CallbackManagerForToolRun] = None,
-    ) -> str:
-        """Generate visualizations from the last SQL query result."""
-        if _is_strict_dashboard_mode(workspace_state) and not _is_plan_approved(workspace_state):
-            return _dashboard_plan_gate_message()
-        if _is_strict_dashboard_mode(workspace_state):
-            return (
-                "Dashboard mode uses structured chart specs instead of generated chart code. "
-                "Skip generate_chart_config and pass the approved chart bindings directly to generate_dashboard."
-            )
-        try:
-            db_manager.require_query_before_chart()
-            db_manager.require_chart_budget()
-        except ValueError as exc:
-            return str(exc)
-
-        df_context = db_manager.get_limited_result()
-        if df_context is None:
-            return "No query results available to visualize."
-
-        charts_dir = workspace_state.root_path / "charts"
-        safe_title = _safe_slug(chart_title, "chart")
-        before_snapshot = _snapshot_workspace(workspace_state.root_path)
-
-        try:
-            sanitized_code = sanitize_python_code(python_code)
-        except ValueError as exc:
-            return _chart_payload(
-                chart_title,
-                status="error",
-                message=str(exc),
-                row_count=len(df_context),
-                error_type="sanitization_error",
-                available_columns=list(df_context.columns),
-            )
-
-        execution_result = _run_chart_code_in_subprocess(
-            sanitized_code,
-            chart_title,
-            df_context,
-        )
-        if execution_result.get("status") != "success":
-            after_snapshot = _snapshot_workspace(workspace_state.root_path)
-            _cleanup_new_files(workspace_state.root_path, before_snapshot, after_snapshot)
-            return _chart_payload(
-                chart_title,
-                status="error",
-                message=execution_result.get("message") or "Chart execution failed.",
-                row_count=len(df_context),
-                error_type=execution_result.get("error_type") or "execution_error",
-                available_columns=execution_result.get("available_columns"),
-            )
-
-        chart_config = execution_result.get("chart_config")
-        plotly_payload = execution_result.get("plotly_payload")
-
-        after_snapshot = _snapshot_workspace(workspace_state.root_path)
-        artifacts = _detect_new_files(
-            workspace_state.root_path, before_snapshot, after_snapshot
-        )
-
-        plotly_config_path: Optional[Path] = None
-        if plotly_payload is not None:
-            try:
-                charts_dir.mkdir(exist_ok=True)
-                plotly_config_path = charts_dir / f"{safe_title}.plotly.json"
-                plotly_config_path.write_text(_json_dump(plotly_payload), encoding="utf-8")
-                artifacts.append(
-                    {
-                        "path": _workspace_rel(plotly_config_path, workspace_state.root_path),
-                        "mimeType": ALLOWED_ARTIFACT_EXTENSIONS[".json"],
-                        "size": plotly_config_path.stat().st_size,
-                        }
-                    )
-            except Exception:  # pragma: no cover - best effort persistence
-                logger.warning("Failed to persist Plotly config", exc_info=True)
-
-        if chart_config is None and plotly_payload is None and not artifacts:
-            return _chart_payload(
-                chart_title,
-                status="error",
-                message="No chart_config variable created and no artifacts produced.",
-                row_count=len(df_context),
-                error_type="empty_result",
-                available_columns=list(df_context.columns),
-            )
-
-        artifact_paths = [artifact["path"] for artifact in artifacts]
-        db_manager.record_chart(chart_title, artifact_paths)
-        for artifact in artifacts:
-            db_manager.register_artifact(artifact)
-        _emit_artifacts(callbacks, artifacts)
-
-        return _chart_payload(
-            chart_title,
-            status="success",
-            message="Chart generated successfully.",
-            row_count=len(df_context),
-            chart_config=plotly_payload if plotly_payload is not None else chart_config,
-            plotly_config_path=(
-                _workspace_rel(plotly_config_path, workspace_state.root_path)
-                if plotly_config_path
-                else None
-            ),
-            output_files=artifacts,
-        )
-
-    @tool
-    def generate_summary(
-        summary: Annotated[str, Field(description="The summary of the actions performed")],
-        insights: Annotated[str, Field(description="The insights from the data")],
-        output_path: Annotated[
-            str,
-            Field(description="Optional stable HTML output path such as reports/orders_daily.html."),
-        ] = "",
-        callbacks: Optional[CallbackManagerForToolRun] = None,
-    ) -> str:
-        """Produce a summary of the results and save it as a self-contained HTML report."""
-        if str(workspace_state.context.get("active_skill") or "").strip() == "data/dashboard":
-            return (
-                "The active skill is data/dashboard. Stay on the dashboard path: request plan approval, "
-                "generate the approved charts, then call generate_dashboard."
-            )
-        if db_manager.session.query_count == 0:
-            return "Run at least one SQL query before summarizing the findings."
-        try:
-            db_manager.ensure_single_summary()
-        except ValueError as exc:
-            return str(exc)
-        db_manager.mark_summary_generated()
-        generated_at = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-        materialization_items: List[str] = []
-        query_items: List[str] = []
-        visualization_items: List[str] = []
-
-        if db_manager.session.materialization_history:
-            for item in db_manager.session.materialization_history:
-                materialization_items.append(
-                    "<article class=\"stack-item\">"
-                    f"<div class=\"stack-meta\">Connector {html.escape(item.connector)} • Rows {item.row_count} • Cached {'yes' if item.cached else 'no'}</div>"
-                    f"<p><strong>Parquet:</strong> <code>{html.escape(item.parquet_path)}</code><br />"
-                    f"<strong>Metadata:</strong> <code>{html.escape(item.metadata_path)}</code><br />"
-                    f"<strong>Expires:</strong> {html.escape(item.expires_at or 'n/a')}</p>"
-                    f"<pre><code>{html.escape(item.sql)}</code></pre>"
-                    "</article>"
-                )
-
-        if db_manager.session.query_history:
-            for idx, query_record in enumerate(db_manager.session.query_history, start=1):
-                row_label = (
-                    f"{query_record.row_count}+ rows previewed"
-                    if query_record.truncated
-                    else f"{query_record.row_count} rows"
-                )
-                query_block = [
-                    "<article class=\"stack-item\">",
-                    f"<div class=\"stack-meta\">Query {idx} • {row_label}</div>",
-                    f"<pre><code>{html.escape(query_record.sql)}</code></pre>",
-                ]
-                if not query_record.preview.empty:
-                    query_block.append("<h4>Sample Data</h4>")
-                    query_block.append(query_record.preview.to_html(index=False, border=0))
-                query_block.append("</article>")
-                query_items.append("".join(query_block))
-
-        run_chart_paths: List[str] = []
-        for chart_record in db_manager.session.chart_history:
-            run_chart_paths.extend(chart_record.artifact_paths)
-
-        plotly_json_files = sorted(
-            p for p in (workspace_state.root_path / rel for rel in run_chart_paths if rel.endswith(".plotly.json"))
-            if p.exists()
-        )
-        plotly_html_files = sorted(
-            p for p in (workspace_state.root_path / rel for rel in run_chart_paths if rel.endswith(".plotly.html"))
-            if p.exists()
-        )
-        png_files = sorted(
-            p for p in (workspace_state.root_path / rel for rel in run_chart_paths if rel.endswith(".png"))
-            if p.exists()
-        )
-        if plotly_json_files or plotly_html_files or png_files:
-            for idx, json_path in enumerate(plotly_json_files, start=1):
-                try:
-                    fig_json = json.loads(json_path.read_text(encoding="utf-8"))
-                    script_payload = json.dumps(fig_json)
-                    div_id = f"plotly-json-{idx}"
-                    visualization_items.append(
-                        "<article class=\"stack-item\">"
-                        f"<h3>{_chart_title_from_path(json_path)}</h3>"
-                        f"<div id=\"{div_id}\" class=\"plotly-embed\"></div>"
-                        "<script>"
-                        f"const spec{idx} = {script_payload};"
-                        f"const data{idx} = spec{idx}.data || []; const layout{idx} = spec{idx}.layout || {{}}; const config{idx} = spec{idx}.config || {{}}; const frames{idx} = spec{idx}.frames || undefined;"
-                        f"Plotly.newPlot('{div_id}', data{idx}, layout{idx}, config{idx}).then(() => {{ if (frames{idx} && frames{idx}.length) {{ Plotly.addFrames('{div_id}', frames{idx}); }} }});"
-                        "</script>"
-                        "</article>"
-                    )
-                except Exception:
-                    logger.warning("Failed to embed Plotly JSON %s", json_path, exc_info=True)
-
-            for html_path in plotly_html_files:
-                try:
-                    html_fragment = html_path.read_text(encoding="utf-8")
-                    visualization_items.append(
-                        "<article class=\"stack-item\">"
-                        f"<h3>{_chart_title_from_path(html_path)}</h3>"
-                        f"<div class=\"plotly-embed\">{html_fragment}</div>"
-                        "</article>"
-                    )
-                except Exception:
-                    logger.warning("Failed to embed Plotly HTML %s", html_path, exc_info=True)
-
-            for png_path in png_files:
-                try:
-                    encoded = base64.b64encode(png_path.read_bytes()).decode("utf-8")
-                    visualization_items.append(
-                        "<article class=\"stack-item\">"
-                        f"<h3>{_chart_title_from_path(png_path)}</h3>"
-                        f"<img src=\"data:image/png;base64,{encoded}\" alt=\"{png_path.stem}\" />"
-                        "</article>"
-                    )
-                except Exception:
-                    logger.warning("Failed to embed PNG %s", png_path, exc_info=True)
-        summary_metric_cards = [
-            {
-                "label": "Queries Run",
-                "value": str(len(db_manager.session.query_history)),
-                "meta": "SQL steps captured in this report",
-            },
-            {
-                "label": "Charts Embedded",
-                "value": str(len(visualization_items)),
-                "meta": "Visual artifacts included below",
-            },
-            {
-                "label": "Warehouse Pulls",
-                "value": str(len(db_manager.session.materialization_history)),
-                "meta": "Materializations cached or refreshed",
-            },
-        ]
-        if db_manager.session.query_history:
-            latest_query = db_manager.session.query_history[-1]
-            summary_metric_cards.append(
-                {
-                    "label": "Latest Result",
-                    "value": str(latest_query.row_count),
-                    "meta": "Rows in the final SQL step",
-                }
-            )
-
-        report_html = render_summary_html(
-            title="Data Analysis Report",
-            generated_at=generated_at,
-            summary_html=_markdown_to_html(summary),
-            insights_html=_markdown_to_html(insights),
-            metric_cards=summary_metric_cards,
-            materialization_items=materialization_items,
-            query_items=query_items,
-            visualization_items=visualization_items,
-        )
-
-        try:
-            reports_dir = workspace_state.root_path / "reports"
-            reports_dir.mkdir(exist_ok=True)
-            if output_path.strip():
-                report_path = _resolve_workspace_html_path(
-                    workspace_state.root_path,
-                    output_path,
-                    default_dir="reports",
-                    default_stem="analysis_report",
-                )
-            else:
-                timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
-                report_path = reports_dir / f"analysis_report_{timestamp}.html"
-            report_path.parent.mkdir(parents=True, exist_ok=True)
-            report_path.write_text(report_html, encoding="utf-8")
-
-            artifact = {
-                "path": _workspace_rel(report_path, workspace_state.root_path),
-                "mimeType": ALLOWED_ARTIFACT_EXTENSIONS[".html"],
-                "size": report_path.stat().st_size,
-            }
-            db_manager.register_artifact(artifact)
-            _emit_artifacts(callbacks, [artifact])
-            return (
-                f"Summary: {summary}\n"
-                f"Insights: {insights}\n\n"
-                f"📄 Full report saved to: {_workspace_rel(report_path, workspace_state.root_path)}"
-            )
-        except Exception as exc:  # pragma: no cover - defensive
-            logger.warning("Failed to save HTML report: %s", exc)
-            return f"Summary: {summary}\nInsights: {insights}\n\n(Note: Failed to save report file: {exc})"
-
-    @tool
-    def generate_dashboard(
-        title: Annotated[str, Field(description="The dashboard title shown in the HTML header")],
-        description: Annotated[
-            str,
-            Field(description="Short paragraph describing what this dashboard shows and who it's for"),
-        ],
-        audience: Annotated[
-            str,
-            Field(description="Primary audience for this dashboard, such as Ops leadership or the Board."),
-        ] = "",
-        business_question: Annotated[
-            str,
-            Field(description="The core business question this dashboard answers."),
-        ] = "",
-        decision_questions: Annotated[
-            Optional[List[str]],
-            Field(description="Ordered list of the decisions or questions the dashboard should help answer."),
-        ] = None,
-        layout_template: Annotated[
-            str,
-            Field(description="Named layout template to use for the dashboard, such as executive_driver_dashboard."),
-        ] = "executive_driver_dashboard",
-        hero_eyebrow: Annotated[
-            str,
-            Field(description="Optional short eyebrow label displayed above the dashboard title."),
-        ] = "Interactive Dashboard",
-        headline_takeaway: Annotated[
-            str,
-            Field(description="Optional one-line headline takeaway shown in the hero section."),
-        ] = "",
-        insights: Annotated[
-            Optional[List[str]],
-            Field(description="Optional list of executive takeaways rendered near the top of the dashboard."),
-        ] = None,
-        known_risks: Annotated[
-            Optional[List[str]],
-            Field(description="Optional list of known dashboard caveats or risks."),
-        ] = None,
-        data_quality_notes: Annotated[
-            Optional[List[str]],
-            Field(description="Optional list of data quality or normalization notes."),
-        ] = None,
-        section_titles: Annotated[
-            Optional[List[str]],
-            Field(description="Optional ordered list of section headings, one per chart produced in this run."),
-        ] = None,
-        output_path: Annotated[
-            str,
-            Field(description="Optional stable dashboard package path such as dashboards/orders_overview or dashboards/orders_overview/."),
-        ] = "",
-        dashboard_dataset_path: Annotated[
-            str,
-            Field(
-                description=(
-                    "Optional workspace-relative path to a canonical Parquet/CSV/JSON dataset used for "
-                    "data-backed cross-filtering."
-                )
-            ),
-        ] = "",
-        filter_schema: Annotated[
-            Optional[List[Dict[str, Any]]],
-            Field(
-                description=(
-                    "Optional filter definitions. Each item may include id, field, label, type "
-                    "(categorical/date/datetime/numeric), multi, options, presets, and applies_to."
-                )
-            ),
-        ] = None,
-        chart_bindings: Annotated[
-            Optional[List[Dict[str, Any]]],
-            Field(
-                description=(
-                    "Optional per-chart bindings aligned by chart order or chart_index. Each binding may include "
-                    "chart_type, x_field, y_field, aggregation, series_field, orientation, sort settings, and titles."
-                )
-            ),
-        ] = None,
-        metric_cards: Annotated[
-            Optional[List[Dict[str, Any]]],
-            Field(
-                description=(
-                    "Optional KPI cards for the hero section. Each item may include label, value, and meta."
-                )
-            ),
-        ] = None,
-        callbacks: Optional[CallbackManagerForToolRun] = None,
-    ) -> str:
-        """Assemble all charts and query history from the current run into one HTML dashboard."""
-        if _is_strict_dashboard_mode(workspace_state) and not _is_plan_approved(workspace_state):
-            return _dashboard_plan_gate_message()
-        try:
-            db_manager.ensure_single_dashboard()
-        except ValueError as exc:
-            return str(exc)
-        normalized_bindings = _normalize_chart_bindings(chart_bindings)
-        filter_config = _normalize_filter_schema(filter_schema)
-        has_native_dataset_dashboard = bool(
-            dashboard_dataset_path.strip()
-            and normalized_bindings
-        )
-
-        if db_manager.session.query_count == 0 and not has_native_dataset_dashboard:
-            return "Run at least one SQL query before building a dashboard."
-        strict_dashboard_mode = _is_strict_dashboard_mode(workspace_state)
-        if strict_dashboard_mode and len(normalized_bindings) < STRICT_DASHBOARD_MIN_CHART_COUNT:
-            return (
-                "Dashboard plan incomplete: pass 3 to 5 approved chart bindings to generate_dashboard "
-                "before building the dashboard package."
-            )
-        if strict_dashboard_mode and not _coerce_text_arg(dashboard_dataset_path).strip():
-            return "Dashboard mode requires dashboard_dataset_path so the native renderer and HTML export share the same dataset."
-        if not db_manager.session.chart_history and not has_native_dataset_dashboard:
-            return (
-                "Generate at least one chart before building a static dashboard, or provide "
-                "dashboard_dataset_path, filter_schema, and chart_bindings for a native filterable dashboard."
-            )
-        if strict_dashboard_mode and len(normalized_bindings) > MAX_CHART_COUNT:
-            return (
-                f"Dashboard plan too large: at most {MAX_CHART_COUNT} approved charts are allowed per dashboard package."
-            )
-
-        stable_title = _safe_slug(title, "dashboard")
-        dashboard_dir = _resolve_workspace_dashboard_dir(
-            workspace_state.root_path,
-            output_path,
-            default_dir="dashboards",
-            default_stem=stable_title,
-        )
-        dashboard_rel_path = _workspace_rel(dashboard_dir, workspace_state.root_path)
-        dashboard_meta_path = dashboard_dir / "dashboard.meta.json"
-        dashboard_spec_path = dashboard_dir / "dashboard.spec.json"
-        dashboard_snapshot_path = dashboard_dir / "dashboard.snapshot.html"
-        dashboard_meta_rel_path = _workspace_rel(dashboard_meta_path, workspace_state.root_path)
-        dashboard_spec_rel_path = _workspace_rel(dashboard_spec_path, workspace_state.root_path)
-        dashboard_snapshot_rel_path = _workspace_rel(dashboard_snapshot_path, workspace_state.root_path)
-        dashboard_data_dir = dashboard_dir / "data"
-        dashboard_rows_path = dashboard_data_dir / "dashboard.rows.json"
-        dashboard_rows_rel_path = _workspace_rel(dashboard_rows_path, workspace_state.root_path)
-
-        chart_entries: List[Dict[str, Any]] = []
-        chart_entries_by_index: Dict[int, Dict[str, Any]] = {}
-        chosen_titles = list(section_titles or [])
-
-        for chart_index, chart_record in enumerate(db_manager.session.chart_history, start=1):
-            preferred_path: Optional[Path] = None
-            preferred_kind = ""
-            for suffix, kind in (
-                (".plotly.json", "plotly_json"),
-                (".plotly.html", "plotly_html"),
-                (".png", "png"),
-            ):
-                for rel_path in chart_record.artifact_paths:
-                    if not rel_path.endswith(suffix):
-                        continue
-                    candidate = workspace_state.root_path / rel_path
-                    if candidate.exists():
-                        preferred_path = candidate
-                        preferred_kind = kind
-                        break
-                if preferred_path is not None:
-                    break
-            if preferred_path is None:
-                continue
-            section_title = (
-                chosen_titles[chart_index - 1]
-                if chart_index - 1 < len(chosen_titles)
-                else chart_record.title
-                or _chart_title_from_path(preferred_path)
-            )
-            chart_entry = (
-                {
-                    "chart_index": chart_index,
-                    "title": section_title,
-                    "path": preferred_path,
-                    "kind": preferred_kind,
-                    "binding": normalized_bindings.get(chart_index),
-                }
-            )
-            chart_entries.append(chart_entry)
-            chart_entries_by_index[chart_index] = chart_entry
-
-        dataset_records: List[Dict[str, Any]] = []
-        dataset_schema: List[Dict[str, Any]] = []
-        dataset_format = ""
-        data_filtering_enabled = False
-        if dashboard_dataset_path.strip():
-            dataset_records, dataset_schema, dataset_format = _load_dashboard_dataset(
-                workspace_state.root_path,
-                dashboard_dataset_path,
-            )
-            data_filtering_enabled = bool(dataset_records and filter_config)
-        if strict_dashboard_mode and not dataset_records:
-            return (
-                "Dashboard mode could not load the canonical dataset. "
-                "Confirm dashboard_dataset_path points to a valid workspace CSV, Parquet, or JSON artifact."
-            )
-        if dataset_records and normalized_bindings:
-            validation_errors = _validate_dashboard_runtime_config(
-                filter_config=filter_config,
-                chart_bindings=normalized_bindings,
-                dataset_schema=dataset_schema,
-            )
-            if validation_errors:
-                return "Dashboard chart binding validation failed:\n- " + "\n- ".join(validation_errors)
-
-        normalized_chart_specs = _build_dashboard_chart_specs(normalized_bindings, dataset_schema)
-        normalized_decision_questions = [
-            _coerce_text_arg(item).strip() for item in (decision_questions or []) if _coerce_text_arg(item).strip()
-        ]
-        normalized_insights = [
-            _coerce_text_arg(item).strip() for item in (insights or []) if _coerce_text_arg(item).strip()
-        ]
-        normalized_known_risks = [
-            _coerce_text_arg(item).strip() for item in (known_risks or []) if _coerce_text_arg(item).strip()
-        ]
-        normalized_data_quality_notes = [
-            _coerce_text_arg(item).strip() for item in (data_quality_notes or []) if _coerce_text_arg(item).strip()
-        ]
-        workspace_id = _coerce_text_arg(getattr(workspace_state, "workspace_id", "")).strip() or "workspace"
-        dashboard_id = uuid4().hex
-        dashboard_meta: Dict[str, Any] = {
-            "dashboardId": dashboard_id,
-            "slug": stable_title,
-            "title": title,
-            "specVersion": 2,
-            "runtimeKind": "native",
-            "createdAt": _utc_now().isoformat(),
-            "updatedAt": _utc_now().isoformat(),
-            "status": "generating",
-            "snapshotPath": dashboard_snapshot_rel_path,
-            "specPath": dashboard_spec_rel_path,
-            "metaPath": dashboard_meta_rel_path,
-            "datasetRef": _build_dataset_reference(workspace_state.root_path, dashboard_dataset_path),
-        }
-        dashboard_dir.mkdir(parents=True, exist_ok=True)
-        dashboard_meta_path.write_text(_json_dump(dashboard_meta), encoding="utf-8")
-        _emit_dashboard_artifact(
-            callbacks,
-            {
-                "dashboardPath": dashboard_rel_path,
-                "workspaceId": workspace_id,
-                "dashboardId": dashboard_id,
-                "title": title,
-                "status": "generating",
-            },
-        )
-
-        bound_cards: List[str] = []
-        static_cards: List[str] = []
-        chart_runtime_defs: List[Dict[str, Any]] = []
-        normalized_dashboard_charts: List[Dict[str, Any]] = []
-        dataset_backed_charts_enabled = bool(dataset_records)
-        approved_chart_count = max(len(normalized_chart_specs), len(chart_entries))
-
-        for chart_index in range(1, approved_chart_count + 1):
-            entry = chart_entries_by_index.get(chart_index)
-            chart_spec = normalized_chart_specs.get(chart_index) or {}
-            binding = normalized_bindings.get(chart_index) or {}
-            section_title = (
-                _coerce_text_arg((chart_spec.get("labels") or {}).get("title"))
-                or _coerce_text_arg(binding.get("title"))
-                or (entry or {}).get("title")
-                or (chosen_titles[chart_index - 1] if chart_index - 1 < len(chosen_titles) else "")
-                or f"Chart {chart_index}"
-            )
-
-            if chart_spec:
-                normalized_dashboard_charts.append(
-                    {
-                        **chart_spec,
-                        "title": section_title,
-                    }
-                )
-
-            binding_is_compatible = bool(
-                dataset_backed_charts_enabled
-                and chart_spec
-                and chart_spec.get("liveCapable")
-            )
-            if binding_is_compatible:
-                div_id = f"dashboard-chart-{chart_index}"
-                chart_runtime_defs.append(
-                    {
-                        "chartId": chart_spec.get("chartId") or f"chart_{chart_index}",
-                        "chartIndex": chart_index,
-                        "divId": div_id,
-                        "title": section_title,
-                        "chartType": chart_spec.get("chartType") or "bar",
-                        "dimensionField": chart_spec.get("dimensionField") or "",
-                        "metricField": chart_spec.get("metricField") or "",
-                        "numeratorField": chart_spec.get("numeratorField") or "",
-                        "denominatorField": chart_spec.get("denominatorField") or "",
-                        "xField": chart_spec.get("dimensionField") or "",
-                        "yField": chart_spec.get("metricField") or "",
-                        "aggregation": chart_spec.get("aggregation") or "count",
-                        "seriesField": chart_spec.get("seriesField") or "",
-                        "orientation": chart_spec.get("orientation") or "",
-                        "sortBy": ((chart_spec.get("sort") or {}).get("by")) or "y",
-                        "sortDirection": ((chart_spec.get("sort") or {}).get("direction")) or "desc",
-                        "limit": chart_spec.get("limit") or 0,
-                        "mode": chart_spec.get("mode") or "lines+markers",
-                        "xTitle": ((chart_spec.get("labels") or {}).get("x")) or "",
-                        "yTitle": ((chart_spec.get("labels") or {}).get("y")) or "",
-                    }
-                )
-                chart_descriptor = "Filter-aware chart bound to the canonical dashboard dataset" if data_filtering_enabled else "Interactive chart bound to the canonical dashboard dataset"
-                bound_cards.append(
-                    f"<article class=\"chart-card filter-aware\" data-chart-index=\"{chart_index}\">"
-                    f"<h3>{html.escape(section_title)}</h3>"
-                    f"<p class=\"chart-meta\">{html.escape(chart_descriptor)}</p>"
-                    f"<div id=\"{div_id}\" class=\"plotly-embed\"></div>"
-                    f"<p class=\"chart-note\" id=\"{div_id}-status\"></p>"
-                    f"</article>"
-                )
-                continue
-
-            if not entry:
-                continue
-
-            chart_path = entry["path"]
-            chart_kind = entry["kind"]
-            static_reason = (
-                _coerce_text_arg(chart_spec.get("staticReason") or binding.get("static_reason"))
-                or "Static appendix artifact because this chart is not bound to the dashboard dataset."
-            )
-            try:
-                if chart_kind == "plotly_json":
-                    fig_json = json.loads(chart_path.read_text(encoding="utf-8"))
-                    script_payload = json.dumps(fig_json)
-                    div_id = f"dashboard-static-chart-{chart_index}"
-                    static_cards.append(
-                        f"<article class=\"chart-card static-card\">"
-                        f"<h3>{html.escape(section_title)}</h3>"
-                        f"<p class=\"chart-meta\">Static appendix chart</p>"
-                        f"<p class=\"chart-note\">{html.escape(static_reason)}</p>"
-                        f"<div id=\"{div_id}\" class=\"plotly-embed\"></div>"
-                        f"<script>"
-                        f"const dashboardStaticSpec{chart_index} = {script_payload};"
-                        f"Plotly.newPlot('{div_id}', dashboardStaticSpec{chart_index}.data || [], dashboardStaticSpec{chart_index}.layout || {{}}, Object.assign({{responsive:true,displayModeBar:false}}, dashboardStaticSpec{chart_index}.config || {{}}));"
-                        f"</script>"
-                        f"</article>"
-                    )
-                elif chart_kind == "plotly_html":
-                    static_cards.append(
-                        f"<article class=\"chart-card static-card\">"
-                        f"<h3>{html.escape(section_title)}</h3>"
-                        f"<p class=\"chart-meta\">Static appendix chart</p>"
-                        f"<p class=\"chart-note\">{html.escape(static_reason)}</p>"
-                        f"{chart_path.read_text(encoding='utf-8')}"
-                        f"</article>"
-                    )
-                elif chart_kind == "png":
-                    encoded = base64.b64encode(chart_path.read_bytes()).decode("utf-8")
-                    static_cards.append(
-                        f"<article class=\"chart-card static-card\">"
-                        f"<h3>{html.escape(section_title)}</h3>"
-                        f"<p class=\"chart-meta\">Static image artifact</p>"
-                        f"<p class=\"chart-note\">{html.escape(static_reason)}</p>"
-                        f"<img class=\"chart-img\" src=\"data:image/png;base64,{encoded}\" alt=\"{html.escape(chart_path.stem)}\" />"
-                        f"</article>"
-                    )
-            except Exception:
-                logger.warning("Failed to embed dashboard artifact %s", chart_path, exc_info=True)
-
-        query_items = []
-        for idx, query_record in enumerate(db_manager.session.query_history, start=1):
-            row_label = (
-                f"{query_record.row_count}+ rows previewed"
-                if query_record.truncated
-                else f"{query_record.row_count} rows"
-            )
-            query_items.append(
-                "<div class=\"query-item\">"
-                f"<div class=\"query-meta\">Query {idx} • {row_label}</div>"
-                f"<pre class=\"query-block\">{html.escape(query_record.sql)}</pre>"
-                "</div>"
-            )
-        generated_at = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-        hero_meta = (
-            f"Charts embedded: {len(bound_cards) + len(static_cards)} • "
-            f"Queries embedded: {len(db_manager.session.query_history)}"
-        )
-        normalized_metric_cards: List[Dict[str, str]] = []
-        for raw_card in metric_cards or []:
-            if not isinstance(raw_card, dict):
-                continue
-            label = _coerce_text_arg(raw_card.get("label")).strip()
-            value = _coerce_text_arg(raw_card.get("value")).strip()
-            meta = _coerce_text_arg(raw_card.get("meta")).strip()
-            if not label or not value:
-                continue
-            normalized_metric_cards.append({"label": label, "value": value, "meta": meta})
-        dashboard_metric_cards = normalized_metric_cards[:3] if normalized_metric_cards else [
-            {
-                "label": "Charts",
-                "value": str(len(chart_runtime_defs) or len(normalized_dashboard_charts)),
-                "meta": "Approved visuals in the dashboard package",
-            },
-            {
-                "label": "Shared Filters",
-                "value": str(len(filter_config)),
-                "meta": "Controls applied across the dashboard dataset",
-            },
-            {
-                "label": "Dataset Rows",
-                "value": str(len(dataset_records)) if dataset_records else "n/a",
-                "meta": "Canonical dataset rows available to the native renderer",
-            },
-        ]
-        validation_error = _validate_dashboard_spec_inputs(
-            strict_dashboard_mode=strict_dashboard_mode,
-            dashboard_dataset_path=dashboard_dataset_path,
-            audience=audience,
-            business_question=business_question,
-            decision_questions=normalized_decision_questions,
-            layout_template=layout_template,
-            metric_cards=dashboard_metric_cards,
-            chart_specs=normalized_chart_specs,
-        )
-        if validation_error:
-            dashboard_meta.update(
-                {
-                    "updatedAt": _utc_now().isoformat(),
-                    "status": "error",
-                }
-            )
-            dashboard_meta_path.write_text(_json_dump(dashboard_meta), encoding="utf-8")
-            _emit_dashboard_artifact(
-                callbacks,
-                {
-                    "dashboardPath": dashboard_rel_path,
-                    "workspaceId": workspace_id,
-                    "dashboardId": dashboard_id,
-                    "title": title,
-                    "status": "error",
-                },
-            )
-            return validation_error
-        filter_panel_html = ""
-        if data_filtering_enabled:
-            filter_panel_html = (
-                "<details class=\"filter-card filter-panel\" open>"
-                "<summary class=\"filter-panel-summary\">"
-                "<div class=\"filter-summary-copy\">"
-                "<strong>Shared data filters</strong>"
-                "<span>Use collapsible controls to update every compatible chart from the embedded dashboard dataset.</span>"
-                "</div>"
-                "</summary>"
-                "<div class=\"filter-panel-body\">"
-                "<div id=\"dashboard-filter-controls\" class=\"filter-controls\"></div>"
-                "<div class=\"filter-actions\">"
-                "<button type=\"button\" id=\"dashboard-apply-filters\" class=\"secondary\">Apply filters</button>"
-                "<button type=\"button\" id=\"dashboard-reset-filters\">Reset filters</button>"
-                "</div>"
-                f"<div class=\"dataset-meta\">Dataset: {html.escape(dashboard_dataset_path)} • Rows embedded: {len(dataset_records)} • Format: {html.escape(dataset_format)}</div>"
-                "</div>"
-                "</details>"
-            )
-        try:
-            html_output = render_dashboard_html(
-                title=title,
-                description=description,
-                hero_eyebrow=hero_eyebrow,
-                audience=audience,
-                business_question=business_question,
-                headline_takeaway=headline_takeaway,
-                generated_at=generated_at,
-                hero_meta=hero_meta,
-                metric_cards=dashboard_metric_cards,
-                filter_panel_html=filter_panel_html,
-                primary_cards=bound_cards or static_cards,
-                appendix_cards=static_cards if bound_cards and static_cards else [],
-                query_items=query_items,
-                decision_questions=normalized_decision_questions,
-                insights=normalized_insights,
-                known_risks=normalized_known_risks,
-                data_quality_notes=normalized_data_quality_notes,
-                layout_template=layout_template,
-                dataset_records=dataset_records,
-                filter_config=filter_config,
-                chart_runtime_defs=chart_runtime_defs,
-                dataset_schema=dataset_schema,
-                highlights_heading="Decision Drivers" if bound_cards else "Charts",
-            )
-        except Exception:
-            dashboard_meta.update(
-                {
-                    "updatedAt": _utc_now().isoformat(),
-                    "status": "error",
-                }
-            )
-            try:
-                dashboard_meta_path.write_text(_json_dump(dashboard_meta), encoding="utf-8")
-            except Exception:
-                logger.warning("Failed to persist dashboard error manifest", exc_info=True)
-            _emit_dashboard_artifact(
-                callbacks,
-                {
-                    "dashboardPath": dashboard_rel_path,
-                    "workspaceId": workspace_id,
-                    "dashboardId": dashboard_id,
-                    "title": title,
-                    "status": "error",
-                },
-            )
-            raise
-        preview_rel_written = ""
-        if dataset_records:
-            dashboard_data_dir.mkdir(parents=True, exist_ok=True)
-            dashboard_rows_path.write_text(_json_dump({"rows": dataset_records}), encoding="utf-8")
-            preview_rel_written = dashboard_rows_rel_path
-        dataset_block = _build_dashboard_dataset_block(
-            workspace_state.root_path,
-            dashboard_dataset_path,
-            dataset_records,
-            preview_rel_written,
-        )
-        dashboard_spec = {
-            "version": 2,
-            "dashboardId": dashboard_id,
-            "runtimeKind": "native",
-            "slug": stable_title,
-            "title": title,
-            "audience": audience,
-            "businessQuestion": business_question,
-            "decisionQuestions": normalized_decision_questions,
-            "description": description,
-            "generatedAt": generated_at,
-            "dashboardPath": dashboard_rel_path,
-            "snapshotPath": dashboard_snapshot_rel_path,
-            "specPath": dashboard_spec_rel_path,
-            "metaPath": dashboard_meta_rel_path,
-            "heroMeta": hero_meta,
-            "highlightsHeading": "Decision Drivers" if bound_cards else "Charts",
-            "hero": {
-                "eyebrow": hero_eyebrow,
-                "description": description,
-                "headlineTakeaway": headline_takeaway,
-                "kpis": dashboard_metric_cards,
-            },
-            "filters": filter_config,
-            "charts": normalized_dashboard_charts,
-            "chartRuntimeDefs": chart_runtime_defs,
-            "metricCards": dashboard_metric_cards,
-            "layout": {
-                "template": layout_template,
-                "gridColumns": 2,
-                "maxFirstScreenVisuals": 5,
-            },
-            "insights": normalized_insights,
-            "knownRisks": normalized_known_risks,
-            "dataQualityNotes": normalized_data_quality_notes,
-            "dataset": dataset_block,
-            "datasetRef": dataset_block,
-            "datasetSchema": dataset_schema,
-            "fallbackMode": "read_only_html",
-        }
-
-        try:
-            dashboard_dir.mkdir(parents=True, exist_ok=True)
-            dashboard_spec_path.write_text(_json_dump(dashboard_spec), encoding="utf-8")
-            dashboard_snapshot_path.write_text(html_output, encoding="utf-8")
-            dashboard_meta.update(
-                {
-                    "updatedAt": _utc_now().isoformat(),
-                    "status": "ready",
-                    "runtimeKind": "native",
-                    "specVersion": 2,
-                    "datasetRef": dashboard_spec.get("dataset") or {},
-                }
-            )
-            dashboard_meta_path.write_text(_json_dump(dashboard_meta), encoding="utf-8")
-            db_manager.mark_dashboard_generated()
-            artifacts_out: List[Dict[str, Any]] = [
-                {
-                    "path": dashboard_snapshot_rel_path,
-                    "mimeType": ALLOWED_ARTIFACT_EXTENSIONS[".html"],
-                    "size": dashboard_snapshot_path.stat().st_size,
-                }
-            ]
-            if preview_rel_written:
-                artifacts_out.append(
-                    {
-                        "path": preview_rel_written,
-                        "mimeType": ALLOWED_ARTIFACT_EXTENSIONS[".json"],
-                        "size": dashboard_rows_path.stat().st_size,
-                    }
-                )
-            for art in artifacts_out:
-                db_manager.register_artifact(art)
-            _emit_artifacts(callbacks, artifacts_out)
-            _emit_dashboard_artifact(
-                callbacks,
-                {
-                    "dashboardPath": dashboard_rel_path,
-                    "workspaceId": workspace_id,
-                    "dashboardId": dashboard_id,
-                    "title": title,
-                    "status": "ready",
-                },
-            )
-            return f"Dashboard package saved to: {dashboard_rel_path}"
-        except Exception as exc:  # pragma: no cover - defensive
-            dashboard_meta.update(
-                {
-                    "updatedAt": _utc_now().isoformat(),
-                    "status": "error",
-                }
-            )
-            try:
-                dashboard_dir.mkdir(parents=True, exist_ok=True)
-                dashboard_meta_path.write_text(_json_dump(dashboard_meta), encoding="utf-8")
-            except Exception:
-                logger.warning("Failed to persist dashboard error manifest", exc_info=True)
-            _emit_dashboard_artifact(
-                callbacks,
-                {
-                    "dashboardPath": dashboard_rel_path,
-                    "workspaceId": workspace_id,
-                    "dashboardId": dashboard_id,
-                    "title": title,
-                    "status": "error",
-                },
-            )
-            logger.warning("Failed to save dashboard HTML: %s", exc)
-            return f"Failed to save dashboard file: {exc}"
-
-    return [
-        get_table_schema,
-        run_sql_query,
-        materialize_bigquery_to_parquet,
-        generate_chart_config,
-        generate_summary,
-        generate_dashboard,
-    ]
+from helpudoc_agent.tools.data.dashboard_tools import _build_dashboard_chart_specs
+from helpudoc_agent.tools.data.duckdb_manager import DuckDBManager
+from helpudoc_agent.tools.data.factory import build_data_agent_tools
+from helpudoc_agent.tools.data.formatting import _format_sample_value
+from helpudoc_agent.tools.data.renderers.dashboard_snapshot import render_dashboard_html
+from helpudoc_agent.tools.data.renderers.html_summary import render_summary_html
+from helpudoc_agent.tools.data.workspace_files import _snapshot_workspace
+
+__all__ = [
+    "build_data_agent_tools",
+    "DuckDBManager",
+    "CHART_EXECUTION_TIMEOUT_SECONDS",
+    "MAX_CHART_COUNT",
+    "MAX_QUERY_COUNT",
+    "_build_dashboard_chart_specs",
+    "_format_sample_value",
+    "_snapshot_workspace",
+    "render_dashboard_html",
+    "render_summary_html",
+    "run_bigquery_query",
+]
