+++ agent/helpudoc_agent/tools/data/__init__.py
@@ -1,0 +1,7 @@
+"""DuckDB-backed data analysis tools (query, charts, dashboards).
+
+Import concrete modules directly, e.g. ``helpudoc_agent.tools.data.factory``,
+to avoid heavy import side effects at package load time.
+"""
+
+__all__: list[str] = []
