+++ agent/helpudoc_agent/tools/data/renderers/__init__.py
@@ -1,0 +1,3 @@
+from .dashboard_snapshot import render_dashboard_html
+from .html_summary import render_summary_html
+__all__ = ["render_summary_html", "render_dashboard_html"]
